% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.numeric_safe.R
\name{as.numeric_safe}
\alias{as.numeric_safe}
\title{Convert x to numeric}
\usage{
as.numeric_safe(x, keep_integer = TRUE)
}
\arguments{
\item{x}{A factor, or a character or numeric vector to be converted to a
numeric vector.}

\item{keep_integer}{\code{TRUE} or \code{FALSE}: return input of \link[=typeof]{type} integer
without converting it to double.}
}
\value{
A numeric vector, possibly \code{numeric(0)} or containing \code{NA_real_}.
}
\description{
Convert \code{x} to a vector with \link[=as.numeric]{numeric} \link{mode} in a way that
works if \code{x} is a \link{factor}, a character vector, or already a numeric vector.
}
\details{
The \link[=typeof]{type} of \link[=integer]{integers} is kept \code{integer} if
\code{keep_integer} is \code{TRUE} and is changed to \code{double} if \code{keep_integer} is
\code{FALSE}.

\code{NULL} and zero-length vectors are converted to \code{numeric(0)}, except for
\code{integer(0)} that is kept unchanged if \code{keep_integer} is \code{TRUE} (which is the
default). In contrast to \code{\link[=as.numeric]{as.numeric()}}, logical vectors of length larger
than zero are converted to a vector of \code{NA_real_}, with a warning, .

\code{as.numeric_safe()} uses a suggestion from \code{\link[=factor]{factor()}} and
\href{https://cran.r-project.org/doc/FAQ/R-FAQ.html#How-do-I-convert-factors-to-numeric_003f}{\R FAQ 7.10}
that works if \code{x} is a factor, a character vector or already is a numeric
vector (even though the last example of \code{as.numeric()} comments that this
approach is less efficient for long vectors).

In contrast, the alternative conversions \code{as.numeric(levels(x))[x]} (suggested
in the \code{Warning} of \code{\link[=factor]{factor()}}) and \code{as.numeric(levels(x))[as.integer(x)]}
(suggested in \R \verb{FAQ 7.10} linked to above) \strong{only} work if \code{x} is factor,
see the \code{Examples}.

Furthermore, \code{as.numeric()} \strong{cannot} be used to convert a factor to numeric,
because it returns their underlying numeric integer representation (i.e., the
indices of the factor levels), see the \code{Warnings} in \code{\link[=as.numeric]{as.numeric()}} and
\code{\link[=factor]{factor()}}.
}
\examples{
x_int <- 5:7
x_num <- as.numeric(x_int)
x_char <- as.character(x_int)
x_fact <- as.factor(x_int)

str(as.numeric_safe(x_int, keep_integer = TRUE))
str(as.numeric_safe(x_int, keep_integer = FALSE))

str_as_num_safe <- function(x) {str(as.numeric_safe(x))}
str_as_num_fact <- function(x) {str(as.numeric(levels(x))[x])}
str_as_num_7.10 <- function(x) {str(as.numeric(levels(x))[as.integer(x)])}
str_as_num_base <- function(x) {str(as.numeric(x))}

# as.numeric_safe() works irrespective the type of x
str_as_num_safe(x_int)  # correct
str_as_num_safe(x_num)  # correct
str_as_num_safe(x_char) # correct
str_as_num_safe(x_fact) # correct

# The 'more efficient' suggestions in `help(factor)` and
# R FAQ 7.10 *only* work for factors.
str_as_num_fact(x_int)  # clearly wrong
str_as_num_7.10(x_int)  # clearly wrong
str_as_num_fact(x_num)  # clearly wrong
str_as_num_7.10(x_num)  # clearly wrong
str_as_num_fact(x_char) # clearly wrong
str_as_num_7.10(x_char) # clearly wrong
str_as_num_fact(x_fact) # correct
str_as_num_7.10(x_fact) # correct

# as.numeric() gives the *indices* of the factor levels
# for factors
str_as_num_base(x_int)  # correct
str_as_num_base(x_num)  # correct
str_as_num_base(x_char) # correct
str_as_num_base(x_fact) # wrong

}
\seealso{
\code{\link[=formatC]{formatC()}} and \link[=options]{option} \code{scipen} on formatting numbers,
\code{\link[utils:type.convert]{utils::type.convert()}}

Other functions to convert types:
\code{\link{reexports}},
\code{\link[=reorder_levels]{reorder_levels()}},
\code{\link[=vect_to_char]{vect_to_char()}}

Other functions to modify character vectors:
\code{\link{reexports}},
\code{\link[=replace_vals]{replace_vals()}},
\code{\link[=signal_text]{signal_text()}},
\code{\link[=unpaste_unquote]{unpaste_unquote()}},
\code{\link[=vect_to_char]{vect_to_char()}},
\code{\link[=wrap_text]{wrap_text()}}

Other functions to modify factors:
\code{\link{reexports}},
\code{\link[=reorder_levels]{reorder_levels()}},
\code{\link[=replace_vals]{replace_vals()}},
\code{\link[=round_levels]{round_levels()}},
\code{\link[=vect_to_char]{vect_to_char()}}
}
\concept{functions to convert types}
\concept{functions to modify character vectors}
\concept{functions to modify factors}
