% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_os_is_windows.R
\name{check_os_is_windows}
\alias{check_os_is_windows}
\title{Check if the operating system is Windows}
\usage{
check_os_is_windows(action = c("warn", "message", "quiet"))
}
\arguments{
\item{action}{Character string indicating the action if the operating system
is \emph{not} Windows.}
}
\value{
\code{TRUE} or \code{FALSE}, indicating if the operating system is Windows, returned
\link[=invisible]{invisibly}.
}
\description{
Check if the operating system is Windows
}
