% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dir.R
\name{create_dir}
\alias{create_dir}
\title{Create a directory}
\usage{
create_dir(dir = fs::path(".", "output"), add_date = TRUE)
}
\arguments{
\item{dir}{\link[checkinput:is_character]{Character string}
containing a \link[checkinput:is_path]{valid path} to a directory that should
be created if it does not yet exist. \code{\link[fs:path]{fs::path()}} adds file separators and
the dot (\code{"."}) indicates the \link[=getwd]{working directory}, such that by
default a subdirectory with the current date in the \link[=strftime]{format}
\code{YYYY_mm_dd} in directory \code{output} below the working directory is created.}

\item{add_date}{\code{TRUE} or \code{FALSE}: create a subdirectory in \code{dir} with the
current date in the \link[=strftime]{format} \code{YYYY_mm_dd}?}
}
\value{
A character string with the \link[fs:path_abs]{absolute normalized} path to the
requested directory, returned \link[=invisible]{invisibly}. An error is thrown if
the attempt to create a directory fails, for example because \code{dir} points to
an existing file instead of to an directory.
}
\description{
Create a directory if it does not yet exist.
}
\details{
The \link[fs:path_abs]{absolute, normalised} path is returned such that the
returned path still works if the \link[=getwd]{working directory} changes. On
case-insensitive file systems, normalization also adjusts the case to match
names of directories that are already present (see the \code{Examples}).
}
\section{Side effects}{

The directory indicated by the returned path is \link[fs:dir_create]{created}
if it does not yet exist.
}

\examples{
# Use a temporary subdirectory to not write in the user's directory
my_tempdir <- create_tempdir(pattern = "examplecreatedir")

# Create directory 'dir_one' inside this temporary directory
res_dir_one <- create_dir(dir = fs::path(my_tempdir, "dir_one"),
                          add_date = FALSE)
fs::dir_exists(res_dir_one) # TRUE

# An attempt to create a directory that already exists does not change any
# directory and the same directory is returned.
res_dir_one_v2 <- create_dir(dir = fs::path(my_tempdir, "dir_one"),
                             add_date = FALSE)
identical(res_dir_one, res_dir_one_v2) # TRUE

# On case-insensitive file systems, the directory 'res_dir_ONE' is the same as
# 'res_dir_one'. On case-sensitive file systems it differs in case from
# 'res_dir_one'.
res_dir_ONE <- create_dir(dir = fs::path(my_tempdir, "dir_ONE"),
                          add_date = FALSE)
identical(res_dir_one, res_dir_ONE)

# Create directory 'dir_two' with a subdirectory containing the current date
res_dir_date <- create_dir(dir = fs::path(my_tempdir, "dir_date"),
                           add_date = TRUE)
fs::dir_exists(res_dir_date) # TRUE

# Cleaning up
unlink(my_tempdir, recursive = TRUE)
rm(my_tempdir, res_dir_one, res_dir_one_v2, res_dir_date, res_dir_ONE)

}
\seealso{
\code{\link[checkinput:is_path]{checkinput::is_path()}} to check if a path is valid, with a \verb{Note on paths}
and extensive references about file paths and directories;
\code{\link[=create_tempdir]{create_tempdir()}} for a safe way to create temporary directories;
\code{\link[=create_file_path]{create_file_path()}} to create a file path and creating the indicated
directory if it does not yet exist;
\code{\link[fs:dir_exists]{fs::dir_exists()}} and \code{\link[fs:dir_create]{fs::dir_create()}} (with its base-equivalent
\code{\link[=dir.create]{dir.create()}}) used by this function;

Other functions to handle paths and directories:
\code{\link[=create_file_path]{create_file_path()}},
\code{\link[=create_tempdir]{create_tempdir()}},
\code{\link[=get_file_path]{get_file_path()}}
}
\concept{functions to handle paths and directories}
