% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_file_path.R
\name{create_file_path}
\alias{create_file_path}
\title{Create a file path}
\usage{
create_file_path(
  filename,
  format_stamp = "\%Y_\%m_\%d_\%H_\%M_\%S",
  dir = fs::path(".", "output"),
  add_date = TRUE
)
}
\arguments{
\item{filename}{\link[checkinput:is_character]{character string} with the file
name, including the file
extension like \code{.csv} or \code{.txt}. Should adhere to the restrictions described
in \code{\link[checkinput:is_path]{checkinput::is_path()}}.}

\item{format_stamp}{\link[checkinput:is_character]{character string} indicating
the \link[=strftime]{format} of
the stamp to be added in front of the file name. No stamp is added if
\code{format_stamp} is an empty string (i.e., \code{""}). The formatted stamp is
treated as part of the filename, such that the same restrictions apply, see
\code{Details}.}

\item{dir}{\link[checkinput:is_character]{Character string}
containing a \link[checkinput:is_path]{valid path} to a directory that should
be created if it does not yet exist. \code{\link[fs:path]{fs::path()}} adds file separators and
the dot (\code{"."}) indicates the \link[=getwd]{working directory}, such that by
default a subdirectory with the current date in the \link[=strftime]{format}
\code{YYYY_mm_dd} in directory \code{output} below the working directory is created.}

\item{add_date}{\code{TRUE} or \code{FALSE}: create a subdirectory in \code{dir} with the
current date in the \link[=strftime]{format} \code{YYYY_mm_dd}?}
}
\value{
The created \link[fs:path_abs]{absolute normalized} file path, returned
\link[=invisible]{invisibly}.
}
\description{
Create a file path, creating the indicated directory if it does not yet exist.
}
\details{
\code{filename} \strong{should} contain a file extension (i.e., a dot followed by any
character until the end of the file name) and should \strong{not} contain slashes
or backslashes: use \code{dir} to indicate subdirectories.

The \link[fs:path_abs]{absolute normalised} path is returned such that the
returned path still works if the \link[=getwd]{working directory} changes.

A warning is issued if the \strong{file} indicated by the returned path already
exists. To prevent this when creating files in quick succession, use \code{"\%OSn"}
as part of \code{format_stamp} to create precise stamps by truncating seconds to
\verb{0 <= n <= 6} decimal places, see \code{\link[=strftime]{strftime()}} for details.

\code{create_file_path()} does \strong{not} create the file: use
\code{fs::file_create(create_file_path(filename = "<filename>", ...))} or
\code{file.create(create_file_path(filename = "<filename>", ...))} to do so.

Calls where date nor time stamps are added (e.g.,
\code{create_file_path(filename = "<filename>", format_stamp = "", dir = "<dir>", add_date = FALSE)})
can be replaced by \code{fs::path(fs::dir_create("<dir>"), "<filename>")} if it is
fine to \strong{not} get a warning if the file already exists.
}
\section{Side effects}{

The directory indicated by the returned path is \link[fs:dir_create]{created}
if it does not yet exist.
}

\examples{
# Use a temporary directory to not write in the user's directory
my_tempdir <- create_tempdir(prefix = "examplecreatefilepath")

(create_file_path(filename = "abc.txt", format_stamp = "",
                  dir = my_tempdir, add_date = TRUE))
(create_file_path(filename = "abc.txt", format_stamp = "\%d_\%m_\%Y",
                  dir = my_tempdir, add_date = TRUE))
(create_file_path(filename = "def.html", format_stamp = "",
                  dir = my_tempdir, add_date = FALSE))
(create_file_path(filename = "def.html", format_stamp = "\%d_\%m_\%Y",
                  dir = my_tempdir, add_date = FALSE))
(create_file_path(filename = "abc.txt", format_stamp = "",
                  dir = fs::path(my_tempdir, "subdir"), add_date = TRUE))
(create_file_path(filename = "abc.txt", format_stamp = "\%d_\%m_\%Y",
                  dir = fs::path(my_tempdir, "subdir"), add_date = TRUE))
(create_file_path(filename = "def.html", format_stamp = "",
                  dir = fs::path(my_tempdir, "subdir"), add_date = FALSE))
(create_file_path(filename = "def.html", format_stamp = "\%d_\%m_\%Y",
                  dir = fs::path(my_tempdir, "subdir"), add_date = FALSE))

# Cleaning up
unlink(x = my_tempdir, recursive = TRUE)
rm(my_tempdir)

}
\seealso{
\code{\link[checkinput:is_path]{checkinput::is_path()}} to check if a path is valid, with a \verb{Note on paths}
and extensive references about file paths and directories;
\code{\link[=get_file_path]{get_file_path()}} to check if a file exists and is a unique match to a pattern;
\code{\link[=create_dir]{create_dir()}} (used by this function) to create a directory if it does not
yet exist

Other functions to handle paths and directories:
\code{\link[=create_dir]{create_dir()}},
\code{\link[=create_tempdir]{create_tempdir()}},
\code{\link[=get_file_path]{get_file_path()}}
}
\concept{functions to handle paths and directories}
