% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paths.R
\name{get_paths}
\alias{get_paths}
\title{Get paths to directories}
\usage{
get_paths()
}
\value{
A character vector containing the same elements as \code{\link[=.libPaths]{.libPaths()}}, possibly in
a different order.
}
\description{
Wrapper around \code{\link[=.libPaths]{.libPaths()}} to get the paths \R knows about, with paths
containing the same \R \link[=getRversion]{version number} (\code{x.y.z}) as the
currently running \R session given before other paths.
}
\section{Programming note}{

See \code{InstallPkgs::prepare_install()} where \code{get_paths()} is checked and used.
}

\examples{
.libPaths()
get_paths() # The same elements as above, possibly in a different order
getRversion() # The currently used R version

}
\seealso{
Other functions to check paths and create directories: 
\code{\link{check_file}()},
\code{\link{create_dir}()}
}
\concept{functions to check paths and create directories}
