% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head_tail.R
\name{head_tail}
\alias{head_tail}
\title{Show first and last part of an object}
\usage{
head_tail(x, n = 3L)
}
\arguments{
\item{x}{object of which the first and last part should be returned.}

\item{n}{\link[checkinput:make_natural]{natural} number (default \code{3L}) indicating
the number of elements or rows to return as \code{head} and \code{tail}.}
}
\value{
The \code{n} first and last elements or rows of \code{x}.
}
\description{
Show first and last part of an object
}
\examples{
x <- letters[1:10]
names(x) <- LETTERS[1:10]
x
head_tail(x)
head_tail(x, n = 4)
head_tail(x, n = 40)
try(head_tail(x, n = 0))
try(head_tail(x, n = 4.1))

df <- head_tail(matrix(data = 1:40, ncol = 4,
                       dimnames = list(LETTERS[1:10], letters[1:4])))
head_tail(df)

}
\seealso{
\code{\link[utils:head]{utils::head()}} \code{\link[utils:tail]{utils::tail()}}
}
