% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_file.R
\name{is_file}
\alias{is_file}
\title{Check if a pattern matches a file}
\usage{
is_file(pattern, ignore_case = TRUE, quietly = FALSE)
}
\arguments{
\item{pattern}{Character string containing a \link[base:regex]{regular expression}
to match to file names in the \link[=getwd]{working directory}.}

\item{ignore_case}{\code{TRUE} or \code{FALSE}: should matching file names be
case-insensitive?}

\item{quietly}{\code{TRUE} or \code{FALSE}: should the message with the found file name
be suppressed?}
}
\value{
\code{TRUE} or \code{FALSE} indicating if \code{pattern} matches a file name. A
warning is issued if \code{pattern} matches multiple file names. An error is
thrown if \code{pattern} does not match any file names.
}
\description{
Check if a pattern matches a file that is present in the working directory.
}
\examples{
is_file(pattern <- "AMESP") # TRUE # (file 'NAMESPACE')
is_file(pattern <- "amesp", ignore_case = TRUE) # TRUE # (file 'NAMESPACE')
is_file(pattern <- "amesp", ignore_case = FALSE) # Error reporting presence of case-insensitive match 'NAMESPACE'
is_file(pattern <- "abc", ignore_case = FALSE) # Error reporting no match found.
is_file(pattern <- "ICEN") # TRUE # Reporting multiple matches.

}
\seealso{
\code{\link[=check_file]{check_file()}} \code{\link[=create_dir]{create_dir()}} \code{\link[checkinput:all_characters]{checkinput::is_character()}}
}
