% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_filename.R
\name{is_filename}
\alias{is_filename}
\title{Check that \code{x} is a valid filename}
\usage{
is_filename(filename)
}
\arguments{
\item{filename}{\link[checkinput:is_character]{character string} with the
filename, including the file extension.}
}
\value{
\code{TRUE}: an error occurs if \code{filename} is not a valid filename.
}
\description{
Check that \code{x} is a filename that includes an extension and is likely to be a
valid filename.
}
\details{
This function puts some restrictions on filenames because it is used to check
for valid filenames before creating a file:
\itemize{
\item \code{filename} should include a file extension. The compression extensions
\code{".gz"}, \code{".bz2"} or \code{".xz"} are \strong{not} considered file extensions but are
allowed to be present.
\item \code{filename} should \strong{not} contain any of the characters \verb{"}, \code{*}, \code{/}, \code{:},
\verb{?}, \verb{\\}, \code{|}, \code{<}, or \code{>}: these are not allowed on Windows.
\item \code{filename} should \strong{not} contain any of the control characters (\code{ASCII}
octal codes 000 through 037 and 177, see \code{help("regex")}).
\verb{?}, \verb{\\}, \code{|}, \code{<}, or \code{>}: these are not allowed on Windows.
\item \code{filename} should \strong{not} start with a space: the space is removed on
Windows.
\item The part of \code{filename} before the extension should \strong{not} end in a space
nor in a period: these are not handled well on Windows.
}

Although not enforced by \code{is_filename()}, it is good practice to avoid the
characters \code{+}, \verb{,}, \verb{;}, \code{=}, \code{[}, \verb{]}, \code{!}, \code{$}, \verb{#}, \code{@}

Also avoid the characters \verb{\{}, \verb{\}}, \code{(}, \verb{)}, \verb{'}, \verb{\%}, \code{&}, \if{html}{\out{<backtick>}}, \code{^}, \code{~}.

Ways to make \code{is_filename()} even stricter:
\itemize{
\item do not allow filenames to start with a hyphen or underline
\item do not allow filenames to end with a hyphen or underline
\item case-insensitive matching to \code{filename} to determine if it exists?
\code{filename} should \strong{not} point to a directory (see \code{get_filename.R},
\code{create_tempdir.R}, \code{create_path.R}).
\item Impose a limit on the length:
https://blog.r-project.org/2023/03/07/path-length-limit-on-windows/
\item See also function \code{progutils::create_path()} and packages \code{fs} and \code{usethis}.
}

In contrast to functions from \code{checkinput}, \code{is_filename} will produce an
error if \code{filename} is not a valid filename.
}
\section{References}{

\itemize{
\item https://learn.microsoft.com/en-us/windows/win32/fileio/naming-a-file
\item https://en.wikipedia.org/wiki/Comparison_of_file_systems#Limits
}
}

\examples{
is_filename("abcd.txt")
is_filename("abcd.txt.gz")

try(is_filename("abcd"))
try(is_filename("abcd.gz"))
try(is_filename("ab|cd.txt"))

}
\seealso{
Other functions to handle paths and directories:
\code{\link[=create_dir]{create_dir()}},
\code{\link[=create_path]{create_path()}},
\code{\link[=create_tempdir]{create_tempdir()}},
\code{\link[=file_path_no_ext]{file_path_no_ext()}},
\code{\link[=get_filename]{get_filename()}},
\code{\link[=is_path]{is_path()}}
}
\concept{functions to handle paths and directories}
