% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numvect_to_char.R
\name{numvect_to_char}
\alias{numvect_to_char}
\title{Convert a numeric vector to a character string}
\usage{
numvect_to_char(x, signif = 3L, sep = ": ", collapse = ", ", width = Inf)
}
\arguments{
\item{x}{a numeric vector or a non-dataframe list that can be unlisted to
obtain a numeric vector.}

\item{signif}{positive integer of length one indicating the number of
significant digits to round to.}

\item{sep}{Character string of length one used to separate the names and the
values.}

\item{collapse}{character string of length one to collapse values into a
single character string, or \code{NULL} to return each value as an element of a
character vector.}

\item{width}{A positive number (rounded to a whole number) giving the maximum
line width (in characters) after wrapping. Can be \code{Inf} to not wrap text.}
}
\value{
A character string with the values in \code{x}, preserving names and
rounded to \code{signif} digits.
}
\description{
Convert a numeric vector to a character string, optionally preserving names
and rounding the values.
}
\section{Wishlist}{

Write a similar function for dataframes or matrices, with the name of the
relevant columns only once.

Implement customized rounding using \code{signif_custom()}, e.g.,
\code{rounded_vals <- signif_custom(x = x, digits = signif, type = type)}
}

\section{Programming note}{

To get a cross-tabulation of \code{x} into a character string, one could use
\code{paste0(numvect_to_char(c(table(z)), sep = " (", collapse =  "), "), ")")},
see the last \code{Example}.
}

\examples{
x <- 1:3
names(x) <- letters[x]
numvect_to_char(x = x) # "a: 1, b: 2, c: 3"
numvect_to_char(x = unname(x)) # "1, 2, 3"
y <- x
y <- y / 7
numvect_to_char(x = y, signif = 7) # "a: 0.1428571, b: 0.2857143, c: 0.4285714"
numvect_to_char(x = y, signif = 2, sep = " = ", collapse = " and ", width = 15)
# "a = 0.14 and b\n= 0.29 and c =\n0.43"

# Showing the use of numvect_to_char to get a frequency table
x <- 1:10
names(x) <- letters[x]
y <- 5:15
names(y) <- letters[y]
z <- c(x, y)
paste0(numvect_to_char(c(table(z)), sep = " (", collapse =  "), "), ")")

}
\seealso{
\code{\link[base:toString]{base::toString()}} which can be used as \code{toString(x)} or
\code{toString(signif_custom(x, ...))} if \code{x} is unnamed or names can be removed.

Other functions to modify character vectors: 
\code{\link{as.numeric_safe}()},
\code{\link{paste_quoted}()},
\code{\link{wrap_text}()}
}
\concept{functions to modify character vectors}
