% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_quoted.R
\name{paste_quoted}
\alias{paste_quoted}
\title{Concatenate text to a single string of quoted elements.}
\usage{
paste_quoted(x, ...)
}
\arguments{
\item{x}{Vector to be converted to a single character string.}

\item{...}{Not used. Present as a programming tool, see \code{Details}.}
}
\value{
A character string consisting of the elements of \code{x} surrounded by
quotes (i.e., \verb{'}), separated by commas (i.e., \verb{,}).
}
\description{
Concatenate text to a single string of quoted elements, printing \code{NULL} and
\code{character(0)} as \code{"'NULL'"}.
}
\details{
The use of ellipsis (\code{...}) as second argument makes it possible to detect if
arguments other than \code{x} are not named. Then a warning is raised because then
\code{x} was probably accidentally not concatenated. For example, the call
\code{paste_quoted("a", "b")} would return \code{"bab"} if no ellipses were present
(using \code{'b'} as argument \code{surround}), whereas now \code{"'a'"} is returned, with a
warning that the probably intended call is \code{paste_quoted(c("a", "b"))}.
}
\note{
\code{NULL} and \code{character(0)} are printed as \code{"'NULL'"}. Other zero-length
objects are printed as \code{"''"}.

A warning is issued if \code{x} has names or if unnamed arguments (other
than \code{x}) are present, see the \code{Wishlist}.
}
\section{To do}{

See also \code{\link[=dQuote]{dQuote()}}, \code{\link[=sQuote]{sQuote()}} and \code{\link[=Quotes]{Quotes()}}

Is there a way to check that arguments are not formal arguments without using
the ellipsis argument? See (1) \code{test_function()} for a more elaborate code to
warn about arguments that are not formal arguments of the called function;
(2) \code{\link[=chkDots]{chkDots()}} with argument \code{allowed} implemented from \R 4.6.0;
(3) \link[methods:dotsMethods]{methods::dotsMethods}; (4) functions \code{check_dots_empty()},
\code{check_dots_unnamed()}, and \code{check_dots_used()} in package \code{ellipsis};
(5a) \code{chk_used()} and \code{vld_used()} from package \code{chk};
(5b) \href{https://stackoverflow.com/questions/64001629}{this StackOverflow post};
(6) \href{https://www.rostrum.blog/posts/2024-03-12-eclipse/}{this blog post};
(7) ellipsis.txt in my private progutils.
}

\section{Wishlist}{
 Implement preserving names, e.g., using
\code{paste(names(x), x, sep = ": ", collapse = ", ")} (used in \code{\link[=numvect_to_char]{numvect_to_char()}})
or \code{paste0("'", paste(names(x), x, sep = ": ", collapse = "', '"), "'")}. Add
argument 'use_names = c("numeric", "all", "none") to use that?
}

\examples{
paste_quoted(c(3, 4)) # "'3', '4'"
paste_quoted(NULL) # "'NULL'"
paste_quoted(c(a = 3, b = 4)) # "'3', '4'" # Warns about dropping names.

}
\seealso{
\code{\link[=paste0]{paste0()}} \code{\link[=toString]{toString()}} which can be used instead of
\code{paste(x, collapse = ", ")}

Other functions to modify character vectors: 
\code{\link{as.numeric_safe}()},
\code{\link{numvect_to_char}()},
\code{\link{wrap_text}()}
}
\concept{functions to modify character vectors}
