% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progutils-package.R
\docType{package}
\name{progutils-package}
\alias{progutils}
\alias{progutils-package}
\title{progutils: Programming Utilities}
\description{
Programming Utilities of Jesse Alderliesten.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JesseAlderliesten/progutils}
  \item Report bugs at \url{https://github.com/JesseAlderliesten/progutils/issues}
}

}
\author{
\strong{Maintainer}: Jesse Alderliesten \email{jessealderliesten@hotmail.com} (\href{https://orcid.org/0000-0003-1132-4310}{ORCID})

}
\keyword{internal}
