% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_cols.R
\name{reorder_cols}
\alias{reorder_cols}
\title{Reorder columns}
\usage{
reorder_cols(x, new_order)
}
\arguments{
\item{x}{Object with at least one column. Columns should have unique,
syntactically valid names, see \code{\link[checkinput:all_names]{checkinput::all_names()}}.}

\item{new_order}{Character vector with a length larger than zero containing
unique names in the new order of the columns.}
}
\value{
\code{x} with reordered columns.
}
\description{
Reorder columns
}
\details{
Column names of \code{x} that are missing from \code{new_order} are appended to
\code{new_order}, with a warning.

Values in \code{new_order} that are missing from column names of \code{x} are dropped,
with a warning.
}
\examples{
test_df <- data.frame(a = 1:2, b = 11:12, c = 21:22)
reorder_cols(x = test_df, new_order = c("b", "a", "c"))

# reorder_cols() appends column names that are missing from 'new_order' and
# drops values from 'new_order' that are missing from column names, both with
# a warning.
reorder_cols(x = test_df, new_order = c("b", "a", "d"))

}
\seealso{
\code{\link[=order]{order()}} \code{\link[=sort]{sort()}}

Other functions for reordering: 
\code{\link{reorder_levels}()}
}
\concept{functions for reordering}
