% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_text.R
\name{signal_text}
\alias{signal_text}
\title{Signal a text to the user}
\usage{
signal_text(
  text,
  signal = c("error", "warning", "message", "quiet"),
  origin = character(0)
)
}
\arguments{
\item{text}{\link[checkinput:all_characters]{character vector} with text to be
signalled, coerced to a \link[checkinput:is_character]{character string} by
\code{\link[=vect_to_char]{vect_to_char()}}.}

\item{signal}{\link[checkinput:is_character]{character string} indicating the
type of signal to be used:
\code{"error"} to throw an \link[=stop]{error}, \code{"warning"} to issue a \link{warning},
\code{"message"} to show a \link{message}, or \code{"quiet"} to be quiet.}

\item{origin}{Character string pasted behind \code{text} giving the origin of the
message. Ignored if it is \code{character(0)}.}
}
\value{
\link[checkinput:is_character]{character string} containing \code{text}, with a
phrase indicating the origin of the signal if \code{origin} is not \code{NULL},
returned \link[=invisible]{invisibly}.
}
\description{
Signal a text to the user through an error, warning, or message.
}
\details{
\code{text} is coerced to a character vector using \code{\link[=vect_to_char]{vect_to_char()}}, which treats
zero-length input and input with length larger than one better than
\code{\link[=message]{message()}} etc. that use \code{\link[=paste0]{paste0()}}, see the \code{Examples}. Call
\code{\link[=vect_to_char]{vect_to_char()}} beforehand on \code{text} for finer control of rounding and
wrapping of text, see the \code{Examples}.

\code{text} is signalled through an \link[=stop]{error}, \link{warning}, \link{message}, or
quietly, depending on argument \code{signal}. To make \code{text} available for further
queries when using \code{signal_text()} in another function, add the content of
the signal as an attribute to the returned object, see the last \code{Example}.
}
\examples{
test_text <- c("Some text", "Some other text")
try(signal_text(text = test_text, signal = "error"))
signal_text(text = test_text, signal = "warn")
signal_text(text = test_text, signal = "message")
signal_text(text = test_text, signal = "quiet")

# signal_text() handles input to 'text' with length larger than one more
# sensible than message():
test_numbers <- 11:13
signal_text(text = test_numbers, signal = "message")
message(test_numbers)

# Newlines in 'signal' are respected:
signal_text(text = paste0(test_text, collapse = "\n"), signal = "warn")

# 'vect_to_char()' is used to display the different NA's in the signal
# clearer than standard printing:
signal_text(text = NA_integer_, signal = "message") # prints NA_integer_
message(NA_integer_) # prints NA

# Call 'vect_to_char()' beforehand on 'text' for finer control of rounding
# and wrapping of text:
message(test_numbers / 7)
signal_text(text = test_numbers / 7, signal = "message")
signal_text(text = vect_to_char(x = test_numbers / 7, signif = 4, width = 15,
                                collapse = "\n"),
            signal = "message")

# This example shows how to make the content of the signal available for
# further queries:
signal_negative <- function(x, signal = c("error", "warning", "message", "quiet")) {
  text_signal <- "This is a negative number"
  if(x < 0) {
    progutils::signal_text(text = text_signal, signal = signal)
    attributes(x) <- list(text_signal = text_signal)
  } else {
    # Using '""' as attribute such that
    # grepl(pattern = "<text>", x = attr(x, "text_signal"), fixed = TRUE)
    # returns 'FALSE' if <text> is not found in 'text_signal' (using
    # 'character(0)' as attribute would lead to 'logical(0)' as return).
    attributes(x) <- list(text_signal = "")
  }
  x
}

res_negative <- signal_negative(x = -1, signal = "warning")
res_positive <- signal_negative(x = 1, signal = "warning")
res_negative
res_positive
grepl(pattern = "negative", x = attr(res_negative, "text_signal"),
      ignore.case = FALSE, fixed = TRUE) # TRUE
grepl(pattern = "negative", x = attr(res_positive, "text_signal"),
      ignore.case = FALSE, fixed = TRUE) # FALSE

}
\seealso{
Other functions to modify character vectors:
\code{\link[=as.numeric_safe]{as.numeric_safe()}},
\code{\link{reexports}},
\code{\link[=replace_vals]{replace_vals()}},
\code{\link[=unpaste_unquote]{unpaste_unquote()}},
\code{\link[=vect_to_char]{vect_to_char()}},
\code{\link[=wrap_text]{wrap_text()}}
}
\concept{functions to modify character vectors}
