% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpaste_unquote.R
\name{unpaste_unquote}
\alias{unpaste_unquote}
\title{Split a character string into a vector, removing quotation marks}
\usage{
unpaste_unquote(x, collapse = c(", ", "; "), quotemarks = c("'", "\\""))
}
\arguments{
\item{x}{a \link[checkinput:is_character]{character string}.}

\item{collapse}{\link[checkinput:all_characters]{character vector} with
elements that were used to collapse
values when \code{x} was created, and are now used to \link[=strsplit]{split} \code{x} on.
Can be \code{character(0)} to leave \code{x} as a character string.}

\item{quotemarks}{\link[checkinput:all_characters]{character vector} with
quotation marks to be removed from \code{x}. Can be \code{character(0)} to not remove
any quotation marks.}
}
\value{
\code{x} without the quotation marks in \code{quotemarks}, split into a vector on the
string in \code{collapse}.
}
\description{
Split a character string into a vector, removing quotation marks
}
\details{
\code{unpaste_unquote()} does \strong{not} restore \code{NA}s or zero-length elements to
their original values after removing the quotation marks, such that
\code{unpaste_unquote()} is \strong{not} the exact reverse of \code{\link[=paste_quoted]{paste_quoted()}}. For
example, \code{"'NA_character_'"} becomes \code{"NA_character_"} instead of
\code{NA_character_}, \code{"'NULL'"} becomes \code{"NULL"} instead of \code{NULL}, and
\code{"'character(0)'"} becomes \code{"character(0)"} instead of \code{character(0)}.
}
\examples{
x <- paste_quoted(c("ff", "gG", "HH"))
x
unpaste_unquote(x = x, collapse = ", ", quotemarks = "'")
unpaste_unquote(x = x, collapse = ", ", quotemarks = "\"")
unpaste_unquote(x = x, collapse = character(0), quotemarks = c("'", "\""))
unpaste_unquote(x = x, collapse = character(0), quotemarks = character(0))

}
\seealso{
\code{\link[=paste_quoted]{paste_quoted()}} for the approximate opposite of \code{unpaste_unquote()}.

Other functions to modify character vectors:
\code{\link[=as.numeric_safe]{as.numeric_safe()}},
\code{\link{reexports}},
\code{\link[=replace_vals]{replace_vals()}},
\code{\link[=signal_text]{signal_text()}},
\code{\link[=vect_to_char]{vect_to_char()}},
\code{\link[=wrap_text]{wrap_text()}}
}
\concept{functions to modify character vectors}
