% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_equal.R
\name{are_equal}
\alias{are_equal}
\title{Check that values in two numeric vectors are nearly equal}
\usage{
are_equal(x, y, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{x, y}{Numeric vectors with length larger than zero to compare for
equality. They may contain \link{NA}s or \link{NaN}s, see \code{Value}.}

\item{tol}{A small \link[checkinput:is_positive]{positive} number. Numbers that differ less
than \code{tol} are considered equal.}
}
\value{
A vector with logical values (\code{TRUE}, \code{FALSE} or \code{NA}) indicating if
elements in \code{x} and \code{y} are equal to each other. \link{NA} is returned for
comparisons involving numeric \code{NA} (i.e., \code{NA_integer_} or \code{NA_real_}),
\link{NaN}, or \link[=Inf]{infinite values} with the same sign.
}
\description{
Check element-wise near-equality of numeric vectors while allowing for small
numeric errors to make \code{are_equal()} safer than \code{\link[=Comparison]{==}}.
}
\details{
\code{are_equal()} allows for small numeric errors when comparing numbers. Such
numeric errors can arise because of rounding or representation error. As the
\code{Note} at \code{\link{==}} warns, \code{x == round(x)} does \strong{not} allow for such errors
but tests exact equality.
}
\section{Acknowledgement}{

Code \code{abs(x - y) < tol} was taken from
\href{https://dplyr.tidyverse.org/reference/near.html}{\code{dplyr::near()}}.
}

\examples{
x <- sqrt(2)^2
x == 2 # FALSE
x - 2 # about 4.44e-16
are_equal(x = x, y = 2) # TRUE

are_equal(x = c(2, 3, 3,        NA, Inf),
          y = c(2, 3, 3 + 1e-8, NA, Inf))
are_equal(x = 3, y = c(2, 3, 3 + 1e-8, NA, Inf))

}
\seealso{
\code{\link[checkinput:is_natural]{checkinput::is_natural()}} to check for element-wise near-equality to natural
numbers; \code{\link[=all.equal]{all.equal()}} to check more generally for near-equality;
\code{\link[=identical]{identical()}}, \code{\link[=not_in]{not_in()}} and \code{\link[=match]{match()}} (containing \code{\%in\%} and, from \R
\verb{4.6.0} onwards, \code{'\%notin\%'}) to check for exact (in)equality, with
\code{\link{Comparison}} to do so using binary operators;
\href{https://CRAN.R-project.org/doc/manuals/R-FAQ.html#Why-doesn_0027t-R-think-these-numbers-are-equal_003f}{\R FAQ 7.31}
for background on numerical equality;
the vignette \emph{Type coercion} in package \code{checkinput}:
\code{vignette("Type_Coercion", package = "checkinput")}.

Other functions to check equality:
\code{\link[=check_case]{check_case()}},
\code{\link[=get_file_path]{get_file_path()}},
\code{\link[=not_in]{not_in()}},
\code{\link[=replace_nonalnum]{replace_nonalnum()}},
\code{\link[=replace_vals]{replace_vals()}},
\code{\link[=signif_custom]{signif_custom()}}
}
\concept{functions to check equality}
