% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_equal.R
\name{are_equal}
\alias{are_equal}
\title{Are numbers equal within a tolerance?}
\usage{
are_equal(x, y, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{x, y}{Numeric vectors to compare for equality.}

\item{tol}{A small \link[checkinput:is_number]{positive} number. Numbers with
absolute differences smaller than \code{tol} are considered to be equal.}
}
\value{
A vector with logical values (\code{TRUE}, \code{FALSE} or \code{NA}) indicating if
elements in \code{x} and \code{y} are equal to each other. \link{NA} is returned for
comparisons involving \code{NA}s or \link{NaN}s and for comparisons of
\link[=Inf]{infinite values} with the same sign.
}
\description{
Test element-wise equality of numeric vectors while allowing for small
numeric differences. The tolerance makes \code{are_equal()} safer than
\link[=Comparison]{==} when used on numeric vectors.
}
\section{Acknowledgement}{

Code taken from \code{dplyr::near()}, with added argument checking.
}

\section{Programming note}{

Legacy code contained \code{near_adj()}, a more-elaborate version of \code{are_equal()}
that also worked on \link{data.frame}s and \link[=matrix]{matrices} (with checks for
compatible \link[=dim]{dimensions}) and contained optional type conversion.
}

\examples{
x <- sqrt(2)^2
x == 2 # FALSE
x - 2 # about 4.44e-16
are_equal(x = x, y = 2) # TRUE

are_equal(x = c(2, 3, 3,         NA, Inf),
          y = c(2, 3, 3 + 1e-8, NA, Inf))
are_equal(x = 3, y = c(2, 3, 3 + 1e-8, NA, Inf))

}
\seealso{
\code{\link[=all.equal]{all.equal()}} to compare objects more generally for near-equality,
\link{Comparison} to compare two vectors using binary operators, \code{\link[=match]{match()}} to
compare character vectors.
}
