% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.numeric_safe.R
\name{as.numeric_safe}
\alias{as.numeric_safe}
\title{Convert x to numeric}
\usage{
as.numeric_safe(x)
}
\arguments{
\item{x}{A character, factor, or numeric object to be converted to numeric.}
}
\value{
A numeric object.
}
\description{
Convert x to numeric in a way that works irrespective if \code{x} is character,
factor, or already numeric.
}
\details{
\code{as.numeric()} \emph{cannot} be used to convert a factor to numeric, because that
returns their underlying numeric (integer) representation (i.e., the indices
of the factor levels), see the \code{Warnings} in \code{\link[=as.numeric]{as.numeric()}} and \code{\link[=factor]{factor()}}.

The conversions \code{as.numeric(levels(x))[x]} (suggested in the \code{Warning} of
\code{\link[=factor]{factor()}} and \code{as.numeric(levels(x))[as.integer(x)]} (suggested in
\href{https://cran.r-project.org/doc/FAQ/R-FAQ.html#How-do-I-convert-factors-to-numeric_003f}{\R FAQ 7.10})
\emph{only} work if \code{x} is factor.

\code{as.numeric_safe()} uses the alternative also suggested in \code{factor()} and
\R \verb{FAQ 7.10} that works irrespective if \code{x} is a factor, character or
numeric (even though the last example of \code{as.numeric} comments that it is
less efficient for long vectors).
}
\section{Wishlist}{

Could add argument \code{keep_integer} indicating if input of type \code{\link[=integer]{integer()}}
should be returned as-is, i.e., without converting to numeric type.
}

\examples{
x_int <- 5:7
x_num <- as.numeric(x_int)
x_char <- as.character(x_int)
x_fact <- as.factor(x_int)

str_as_num_safe <- function(x) {str(as.numeric_safe(x))}
str_as_num_base <- function(x) {str(as.numeric(x))}
str_as_num_fact <- function(x) {str(as.numeric(levels(x))[x])}
str_as_num_7.10 <- function(x) {str(as.numeric(levels(x))[as.integer(x)])}

# as.numeric_safe() works irrespective the type of x
str_as_num_safe(x_int)  # num [1:3] 5 6 7 # fine
str_as_num_safe(x_num)  # num [1:3] 5 6 7 # fine
str_as_num_safe(x_char) # num [1:3] 5 6 7 # fine
str_as_num_safe(x_fact) # num [1:3] 5 6 7 # fine

# The 'more efficient' suggestion in help(factor) *only* works for factors.
str_as_num_fact(x_int)  # num [1:3] NA NA NA # wrong, but clearly so.
str_as_num_fact(x_num)  # num [1:3] NA NA NA # wrong, but clearly so.
str_as_num_fact(x_char) # num [1:3] NA NA NA # wrong, but clearly so.
str_as_num_fact(x_fact) # num [1:3] 5 6 7 # fine

# as.numeric() gives the *indices* of the factor levels for factors
str_as_num_base(x_int)  # num [1:3] 5 6 7 # fine
str_as_num_base(x_num)  # num [1:3] 5 6 7 # fine
str_as_num_base(x_char) # num [1:3] 5 6 7 # fine
str_as_num_base(x_fact) # num [1:3] 1 2 3 # wrong!

# The 'more efficient' suggestion in R FAQ 7.10 *only* works for factors.
str_as_num_7.10(x_int)  # num [1:3] NA NA NA # wrong, but clearly so.
str_as_num_7.10(x_num)  # num [1:3] NA NA NA # wrong, but clearly so.
str_as_num_7.10(x_char) # num [1:3] NA NA NA # wrong, but clearly so.
str_as_num_7.10(x_fact) # num [1:3] 5 6 7 # fine

}
\seealso{
Other functions to modify character vectors: 
\code{\link{numvect_to_char}()},
\code{\link{paste_quoted}()},
\code{\link{wrap_text}()}
}
\concept{functions to modify character vectors}
\concept{functions to modify factors}
