% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_case.R
\name{check_case}
\alias{check_case}
\title{Check for values that differ only in their case}
\usage{
check_case(x, signal = c("error", "warning", "message", "quiet"))
}
\arguments{
\item{x}{Character vector to check.}

\item{signal}{\link[checkinput:is_character]{character string} indicating the
type of signal to be used:
\code{"error"} to throw an \link[=stop]{error}, \code{"warning"} to issue a \link{warning},
\code{"message"} to show a \link{message}, or \code{"quiet"} to be quiet.}
}
\value{
A \link[checkinput:all_characters]{character vector} with \link{unique},
\link[=sort]{sorted} values in \code{x} that only differ from each other in their
case, or \code{character(0)} if no such values are present, returned
\link[=invisible]{invisibly}.
}
\description{
Check for values that differ only in their case
}
\details{
Values in \code{x} that only differ from each other in their case lead to a
\link[=signal_text]{signal} as indicated by argument \code{signal}.
}
\section{Notes}{

The sorting order in the result depends on the used \link[=locales]{locale} (see
also the section \code{Details} of \code{\link{Comparison}}), which also affects if uppercase
characters are sorted before or after lowercase characters.
}

\examples{
x <- c("Ee", "Ee", "LL", "Ll")
try(check_case(x = x, signal = "error"))
check_case(x = x, signal = "warning")

}
\seealso{
\code{\link[=tolower]{tolower()}} and \code{\link[=toupper]{toupper()}} to change case.

Other functions to check equality:
\code{\link[=are_equal]{are_equal()}},
\code{\link[=get_file_path]{get_file_path()}},
\code{\link[=not_in]{not_in()}},
\code{\link[=replace_nonalnum]{replace_nonalnum()}},
\code{\link[=replace_vals]{replace_vals()}},
\code{\link[=signif_custom]{signif_custom()}}
}
\concept{functions to check equality}
