% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_case.R
\name{check_case}
\alias{check_case}
\title{Check for values that differ only in their case}
\usage{
check_case(x, signal = c("error", "warning", "message", "quiet"))
}
\arguments{
\item{x}{Character vector to check.}

\item{signal}{Character string indicating the type of signal to be used:
\code{"error"} to throw an \link[=stop]{error}, \code{"warning"} to issue a \link{warning},
\code{"message"} to show a \link{message}, or \code{"quiet"} to be quiet.}
}
\value{
Character vector with \code{\link[=unique]{unique()}} \link[=sort]{sorted} values in \code{x} that only
differ from each other in their case, or \code{character(0)} if no such values are
present. The return is \link{invisible}.
}
\description{
Check for values that differ only in their case
}
\section{Side effects}{

Values in \code{x} that only differ from each other in their case generate an
\link[=stop]{error}, \link{warning}, or \link{message} if \code{signal} is \code{error}, \code{warning}, or
\code{message}, respectively. Values are silently returned if \code{signal} is \code{quiet}.
}

\section{Notes}{

Sort order depends on the used \link[=locales]{locale} (see also the section
\code{Details} of \link{Comparison}), which also affects if uppercase characters are
sorted before or after lowercase characters.
}

\examples{
x <- c("Ee", "Ee", "LL", "Ll")
try(check_case(x = x, signal = "error"))
check_case(x = x, signal = "warning")

}
\seealso{
Other functions to check equality:
\code{\link[=are_equal]{are_equal()}},
\code{\link[=get_filename]{get_filename()}},
\code{\link[=not_in]{not_in()}},
\code{\link[=replace_vals]{replace_vals()}}
}
\concept{functions to check equality}
