% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_file.R
\name{check_file}
\alias{check_file}
\title{Check if a file is present}
\usage{
check_file(pattern, ignore_case = TRUE, quietly = FALSE)
}
\arguments{
\item{pattern}{Character string containing a \link[base:regex]{regular expression}
to match to file names in the \link[=getwd]{working directory}.}

\item{ignore_case}{\code{TRUE} or \code{FALSE}: should matching file names be
case-insensitive?}

\item{quietly}{\code{TRUE} or \code{FALSE}: should the message with the found file name
be suppressed?}
}
\value{
A character string with a file name matching \code{pattern}. The first
match is used if \code{pattern} matches multiple names, with a warning. An error
is thrown if \code{pattern} does not match any names.
}
\description{
Check if a file is present in the working directory.
}
\details{
If \code{ignore_case} is \code{FALSE} and no case-sensitive match is found, it is
checked if a case-insensitive match would be found. If such a match is found,
it is reported in error message.
}
\examples{
check_file(pattern <- "AMESP") # Using file 'NAMESPACE'
check_file(pattern <- "amesp", ignore_case = TRUE) # Using file 'NAMESPACE'
check_file(pattern <- "amesp", ignore_case = FALSE) # Error reporting presence of case-insensitive match 'NAMESPACE'
check_file(pattern <- "abc", ignore_case = FALSE) # Error reporting no match found.
check_file(pattern <- "ICEN") # Multiple matches, using the first.

}
\seealso{
\code{\link[=is_file]{is_file()}} \code{\link[=create_dir]{create_dir()}}
}
