% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dir.R
\name{create_dir}
\alias{create_dir}
\title{Create a directory}
\usage{
create_dir(dir = fs::path(".", "output"), add_date = TRUE)
}
\arguments{
\item{dir}{\link[checkinput:is_character]{Character string}
containing a \link[checkinput:is_path]{valid path} to a directory that should be created if
it does not yet exist. \code{\link[fs:path]{fs::path()}} adds file separators and the dot (\code{"."})
indicates the \link[=getwd]{working directory},
such that by default a subdirectory with the current date in the
\link[=strftime]{format} \code{YYYY_mm_dd} in directory \code{output} below the working
directory is created.}

\item{add_date}{\code{TRUE} or \code{FALSE}: create a subdirectory in \code{dir} with the
current date in the \link[=strftime]{format} \code{YYYY_mm_dd}?}
}
\value{
A character string with the \link[fs:path_abs]{absolute normalized} path to
the requested directory, returned \link[=invisible]{invisibly}. An error is thrown
if the attempt to create a directory fails. This happens if \code{dir} points to
an existing file instead of an directory.
}
\description{
Create a directory if it does not yet exist.
}
\details{
The \link[fs:path_abs]{absolute normalised} path is returned such that the
returned path still works if the \link[=getwd]{working directory} changes. On
case-insensitive file systems (e.g., Windows and macOS), normalization
adjusts the case to match case-insensitive names of directories that are
already present (see the \code{Examples}).
}
\section{Side effects}{

The directory indicated by the returned path is \link[=create_dir]{created} if it
does not yet exist.
}

\examples{
# Use a temporary subdirectory to not write in the user's directory
my_tempdir <- fs::path(tempdir(), "testcreatedir")

# Create directory 'dir_one' inside this temporary directory
res_dir_one <- create_dir(dir = fs::path(my_tempdir, "dir_one"),
                          add_date = FALSE)
fs::dir_exists(res_dir_one) # TRUE

# An attempt to create a directory that already exists does not change any
# directory and the same directory is returned.
res_dir_one_v2 <- create_dir(dir = fs::path(my_tempdir, "dir_one"),
                             add_date = FALSE)
identical(res_dir_one, res_dir_one_v2) # TRUE

# On case-insensitive file systems such as Windows and macOS, adding
# 'dir_ONE' to the directory gives the same result as adding 'dir_one' as
# done above for 'res_dir_one'
res_dir_one_v3 <- create_dir(dir = fs::path(my_tempdir, "dir_ONE"),
                             add_date = FALSE)
# TRUE on Windows and macOS, FALSE on Ubuntu
identical(res_dir_one, res_dir_one_v3)

# Create directory 'dir_two' with a subdirectory containing the current date
res_dir_two <- create_dir(dir = fs::path(my_tempdir, "dir_two"),
                          add_date = TRUE)
fs::dir_exists(res_dir_two) # TRUE

# Cleaning up
unlink(dirname(res_dir_one), recursive = TRUE)
rm(my_tempdir, res_dir_one, res_dir_one_v2, res_dir_two, res_dir_one_v3)

}
\seealso{
\code{\link[checkinput:is_path]{checkinput::is_path()}} to check if a path is valid;
\code{\link[=create_file_path]{create_file_path()}} to create a file path and creating the indicated
directory if it does not yet exist;
\code{\link[=create_tempdir]{create_tempdir()}} for a safe way to create temporary directories;
\code{\link[checkinput:is_path]{checkinput::is_path()}} and references there about file paths and directories;
\code{\link[fs:dir_exists]{fs::dir_exists()}} and \code{\link[fs:dir_create]{fs::dir_create()}} used by this function (and the
base-equivalent of the latter: \code{\link[=dir.create]{dir.create()}});
\code{\link[=get_file_path]{get_file_path()}} to check if a file exists and is a unique match to a
pattern.

Other functions to handle paths and directories:
\code{\link[=create_file_path]{create_file_path()}},
\code{\link[=create_tempdir]{create_tempdir()}},
\code{\link[=get_file_path]{get_file_path()}}
}
\concept{functions to handle paths and directories}
