% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dir.R
\name{create_dir}
\alias{create_dir}
\title{Create a directory}
\usage{
create_dir(dir = file.path(".", "output"), add_date = TRUE)
}
\arguments{
\item{dir}{Non-empty \link[checkinput:is_character]{character string}
containing a \link[=is_path]{valid path} to a directory that should be created if
it does not yet exist. A dot (i.e., \code{"."}) indicates the current working
directory.}

\item{add_date}{\code{TRUE} or \code{FALSE}: create a subdirectory in \code{dir} with the
current date in the \link[=strftime]{format} \code{YYYY_mm_dd}?}
}
\value{
A character string with the absolute \link[=normalizePath]{normalized} path to
the requested directory, returned \link[=invisible]{invisibly}. The
\link[=getwd]{working directory} is returned if an attempt to create a directory
fails, with a warning.
}
\description{
Create a directory if it does not yet exist.
}
\details{
The default \code{dir} is a subdirectory with the current date in the
\link[=strftime]{format} \code{YYYY_mm_dd} in directory \code{output} below the working
directory. \code{\link[=file.path]{file.path()}} ensures the correct (\link[=.Platform]{platform}-dependent)
file separator is used to indicate subdirectories, and \code{"."} indicates the
\link[=getwd]{working directory}.

Several limitations are imposed on \code{dir}, see \code{\link[=is_path]{is_path()}}.

If creating the directory fails, the working directory is returned instead.
This happens if \code{dir} points to an existing file instead of an directory.

The absolute \link[=normalizePath]{normalised} path is returned such that the
returned path still works if the \link[=getwd]{working directory} changes. On
case-insensitive file systems (e.g., Windows and macOS), normalization
adjusts the case to match case-insensitive names of directories that are
already present (see the \code{Examples}). \code{"/"} instead of \code{"\\\\"} is used as
\link[=normalizePath]{winslash} during normalisation, such that the returned path
can be used in Windows' file system.
}
\section{Side effects}{

The requested directory is created if does not yet exist.
}

\section{Programming notes}{

See also:
\itemize{
\item \code{utils::file_test()}
\item \code{fs::is_dir()}
\item \code{fs::is_dir_empty()}.
}
}

\examples{
# Use a temporary subdirectory to not write in the user's directory
my_tempdir <- file.path(tempdir(), "testcreatedir")

# Create directory 'dir_one' inside this temporary directory
res_dir_one <- create_dir(dir = file.path(my_tempdir, "dir_one"),
                          add_date = FALSE)
dir.exists(res_dir_one) # TRUE

# An attempt to create a directory that already exists does not change any
# directory and the same directory is returned.
res_dir_one_v2 <- create_dir(dir = file.path(my_tempdir, "dir_one"),
                             add_date = FALSE)
identical(res_dir_one, res_dir_one_v2) # TRUE

# On case-insensitive file systems such as Windows and macOS, adding
# 'dir_ONE' to the directory gives the same result as adding 'dir_one' as
# done above for 'res_dir_one'
res_dir_one_v3 <- create_dir(dir = file.path(my_tempdir, "dir_ONE"),
                             add_date = FALSE)
# TRUE on Windows and macOS, FALSE on Ubuntu
identical(res_dir_one, res_dir_one_v3)

# Create directory 'dir_two' with a subdirectory containing the current date
res_dir_two <- create_dir(dir = file.path(my_tempdir, "dir_two"),
                          add_date = TRUE)
dir.exists(res_dir_two) # TRUE

# Cleaning up
unlink(dirname(res_dir_one), recursive = TRUE)
rm(my_tempdir, res_dir_one, res_dir_one_v2, res_dir_two, res_dir_one_v3)

}
\seealso{
\code{\link[=create_file_path]{create_file_path()}} to create a file path, \code{\link[=create_tempdir]{create_tempdir()}} for a safe way
to create temporary directories, \code{\link[=is_path]{is_path()}} and references there about file
paths and directories, \code{\link[=dir.exists]{dir.exists()}} and \code{\link[=dir.create]{dir.create()}} used by this
function, \code{\link[=get_file_path]{get_file_path()}} to check if a file exists and is a unique match to
a pattern.

Other functions to handle paths and directories:
\code{\link[=create_file_path]{create_file_path()}},
\code{\link[=create_tempdir]{create_tempdir()}},
\code{\link[=file_path_no_ext]{file_path_no_ext()}},
\code{\link[=get_file_path]{get_file_path()}},
\code{\link[=is_filename]{is_filename()}},
\code{\link[=is_path]{is_path()}}
}
\concept{functions to handle paths and directories}
