% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dir.R
\name{create_dir}
\alias{create_dir}
\title{Create a directory}
\usage{
create_dir(dir = file.path(".", "output"), add_date = TRUE, quietly = FALSE)
}
\arguments{
\item{dir}{Non-empty character string containing the path to the directory,
not ending with a (back)slash. See \code{Details}.}

\item{add_date}{\code{TRUE} or \code{FALSE}: should a subdirectory be created with the
current date?}

\item{quietly}{\code{TRUE} or \code{FALSE}: should the printing of messages indicating
which directory was created be suppressed?}
}
\value{
A character string with the created path, returned
\link[=invisible]{invisibly}. The path will be composed of the elements in \code{dir},
separated with the platform-dependent \link[=file.path]{file separator} and, if
\code{add_date} is \code{TRUE}, a subdirectory with the current date in the format
\code{YYYY_MM_DD}. The \link[=getwd]{working directory} is returned if the attempt to
create a directory fails, with a warning.
}
\description{
Create a directory if it does not yet exist.
}
\details{
The default \code{dir} uses \code{\link[=file.path]{file.path()}} to use the correct (platform-dependent)
path separator when indicating subdirectory \code{output} in the
\link[=getwd]{working directory}. That is not the case if a character string with
(back)slashes (\code{/} or \verb{\\}) is used to indicate subdirectories.
}
\section{Side effects}{

The requested directory is created if does not yet exist.
}

\section{Wishlist}{

Use a temporary directory for the unit tests, to solve the problem that tests
create directories that are left. See the suggestions in test_create_dir.R.
}

\seealso{
\code{\link[=file.path]{file.path()}} and \code{\link[=dir.create]{dir.create()}} that are used in this function.
\code{\link[=check_file]{check_file()}} to check if a file exists.
}
