% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dir.R
\name{create_dir}
\alias{create_dir}
\title{Create a directory}
\usage{
create_dir(dir = file.path(".", "output"), add_date = TRUE, quietly = FALSE)
}
\arguments{
\item{dir}{Non-empty character string containing the path to the directory,
not ending with a (back)slash. See \code{Details}.}

\item{add_date}{\code{TRUE} or \code{FALSE}: should a subdirectory be created with the
current date?}

\item{quietly}{\code{TRUE} or \code{FALSE}: should the printing of messages indicating
which directory was created be suppressed?}
}
\value{
A character string with \link[=normalizePath]{normalized} form of the
created path, returned \link[=invisible]{invisibly}. The path will be composed of
the elements in \code{dir}, separated with the platform-dependent
\link[=file.path]{file separator} and, if \code{add_date} is \code{TRUE}, a subdirectory
with the current date in the format \code{YYYY_MM_DD}. The
\link[=getwd]{working directory} is returned if the attempt to create a directory
fails, with a warning.
}
\description{
Create a directory if it does not yet exist.
}
\details{
The default \code{dir} uses \code{\link[=file.path]{file.path()}} to use the correct (platform-dependent)
path separator when indicating subdirectory \code{output} in the
\link[=getwd]{working directory}. That is not the case if a character string with
(back)slashes (\code{/} or \verb{\\}) is used to indicate subdirectories.
}
\section{Side effects}{

The requested directory is created if does not yet exist.
}

\section{Wishlist}{

Do not allow paths ending in a trailing dot or a trailing space, see
?dir.create().

Add a test for a failed creation.
}

\examples{
# Use a temporary directory to not write in the user's directory
my_tempdir <- tempdir()

# Create directory 'dir_one' inside this temporary directory
res_dir_one <- create_dir(dir = file.path(my_tempdir, "dir_one"),
                          add_date = FALSE)
dir.exists(res_dir_one) # TRUE

# An attempt to create a directory that already exists does not change any
# directory and the same directory is returned. However, the message now
# indicates the directory already exists.
res_dir_one_v2 <- create_dir(dir = file.path(my_tempdir, "dir_one"),
                             add_date = FALSE)
identical(res_dir_one, res_dir_one_v2) # TRUE

# Create directory 'dir_two' with a subdirectory containing the current date
res_dir_two <- create_dir(dir = file.path(my_tempdir, "dir_two"),
                          add_date = TRUE)
dir.exists(res_dir_two) # TRUE

# Cleaning up
unlink(c(res_dir_one, dirname(res_dir_two)), recursive = TRUE)

}
\seealso{
\code{\link[=file.path]{file.path()}} and \code{\link[=dir.create]{dir.create()}} that are used in this function.
\code{\link[=check_file]{check_file()}} to check if a file exists.
}
