% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dir.R
\name{create_dir}
\alias{create_dir}
\title{Create a directory}
\usage{
create_dir(dir = file.path(".", "output"), add_date = TRUE, quietly = FALSE)
}
\arguments{
\item{dir}{Non-empty character string containing the path to a directory.
Case-insensitive matching is used to checking if it already exists.}

\item{add_date}{\code{TRUE} or \code{FALSE}: create a subdirectory with the current
date in the format \code{YYYY_MM_DD}?}

\item{quietly}{\code{TRUE} or \code{FALSE}: suppress the message indicating which
directory was created?}
}
\value{
A character string with the \link[=normalizePath]{normalized} form of the
created path, returned \link[=invisible]{invisibly}. The path will be composed of
\code{dir}, and, if \code{add_date} is \code{TRUE}, a subdirectory with the current date in
the format \code{YYYY_MM_DD}. The \link[=getwd]{working directory} is returned if the
attempt to create a directory fails, with a warning.
}
\description{
Create a directory if it does not yet exist.
}
\details{
The \code{"."} in the default for \code{dir} indicates the \link[=getwd]{working directory}.
Using \code{\link[=file.path]{file.path()}} to create \code{dir} ensures the correct (platform-dependent)
file separator is used to indicate subdirectories.

Several limitations are imposed on \code{dir} to facilitate handling of paths by
Windows, see \code{\link[=dir.create]{dir.create()}}: \code{dir} should not end in a slash, backslash, or
space, which would be removed when the \link[=dir.create]{directory is created},
leading to a mismatch between the created directory and the returned path.
\code{dir} should also not end in a dot. Finally, \code{dir} should not contain the
characters \verb{"}, \code{*}, \verb{?}, \code{|}, \code{<}, or \code{>}.
}
\section{Side effects}{

The requested directory is created if does not yet exist.
}

\section{To do}{

\itemize{
\item What happens if 'dir' points to an existing file instead of to an existing
directory?
\item Why are paths only \link[=normalizePath]{normalized} to use in messages, not
before checking existence?
}
}

\examples{
# Use a temporary directory to not write in the user's directory
my_tempdir <- tempdir()

# Create directory 'dir_one' inside this temporary directory
res_dir_one <- create_dir(dir = file.path(my_tempdir, "dir_one"),
                          add_date = FALSE)
dir.exists(res_dir_one) # TRUE

# An attempt to create a directory that already exists does not change any
# directory and the same directory is returned. However, the message now
# indicates the directory already exists.
res_dir_one_v2 <- create_dir(dir = file.path(my_tempdir, "dir_one"),
                             add_date = FALSE)
identical(res_dir_one, res_dir_one_v2) # TRUE

# Directories are case-insensitive, so adding 'dir_ONE' gives the same result
# as above:
res_dir_one_v3 <- create_dir(dir = file.path(my_tempdir, "dir_ONE"),
                             add_date = FALSE)
identical(res_dir_one, res_dir_one_v3) # TRUE

# Create directory 'dir_two' with a subdirectory containing the current date
res_dir_two <- create_dir(dir = file.path(my_tempdir, "dir_two"),
                          add_date = TRUE)
dir.exists(res_dir_two) # TRUE

# Cleaning up
unlink(c(res_dir_one, dirname(res_dir_two)), recursive = TRUE)

}
\seealso{
\code{\link[=dir.exists]{dir.exists()}} and \code{\link[=dir.create]{dir.create()}} used by this function; \code{\link[=check_file]{check_file()}} to
check if a file exists and is a unique match to a pattern; \code{\link[=file.path]{file.path()}} to
construct the path to a file from components in a platform-independent way;
\code{\link[=normalizePath]{normalizePath()}} to create absolute paths.

\code{fs::path_sanitize()} to \emph{remove} invalid characters from potential paths,
looking for a wider range of invalid characters.

Other functions to check and modify paths and directories: 
\code{\link{check_file}()},
\code{\link{get_paths}()}
}
\concept{functions to check and modify paths and directories}
