% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_file_path.R
\name{create_file_path}
\alias{create_file_path}
\title{Create a file path}
\usage{
create_file_path(
  filename,
  format_stamp = "\%Y_\%m_\%d_\%H_\%M_\%S",
  dir = file.path(".", "output"),
  add_date = TRUE
)
}
\arguments{
\item{filename}{A character string with the file name, including the file
extension like \code{.csv} or \code{.txt}.}

\item{format_stamp}{A character string indicating the \link[=strftime]{format} of
the stamp to be added in front of the file name. No stamp is added if
\code{format_stamp} is an empty string (i.e., \code{""}). The formatted stamp is
treated as part of the filename, such that the same restrictions apply, see
\code{Details}.}

\item{dir}{Non-empty \link[checkinput:is_character]{character string}
containing a \link[=is_path]{valid path} to a directory that should be created if
it does not yet exist. A dot (i.e., \code{"."}) indicates the current working
directory.}

\item{add_date}{\code{TRUE} or \code{FALSE}: create a subdirectory in \code{dir} with the
current date in the \link[=strftime]{format} \code{YYYY_mm_dd}?}
}
\value{
The created file path, returned \link[=invisible]{invisibly}.
}
\description{
Create a file path, creating the indicated directory if it does not yet exist.
}
\details{
\code{filename} should contain a file extension (i.e., a dot followed by
alphanumeric characters until the end of the file name). It should not
contain slashes or backslashes: use \code{dir} to indicate (sub)directories.
Non-alphanumeric characters other than underscores before the file extension
are replaced by dots, with a warning.

The default \code{dir} is a subdirectory with the current date in the
\link[=strftime]{format} \code{YYYY_mm_dd} in directory \code{output} below the working
directory. \code{\link[=file.path]{file.path()}} ensures the correct (\link[=.Platform]{platform}-dependent)
file separator is used to indicate subdirectories, and \code{"."} indicates the
\link[=getwd]{working directory}.

\code{dir} should point to a \link[=is_path]{valid path}. The directory for the
returned path is \link[=create_dir]{created} if it does not yet exist.

The absolute \link[=normalizePath]{normalised} path is returned such that the
returned path still works if the \link[=getwd]{working directory} changes. \code{"/"}
instead of \code{"\\\\"} is used as argument \link[=normalizePath]{winslash} such that
the returned path can be used in Windows' file system.

A warning is issued if the \strong{file} indicated by the returned
path already exists. To prevent this when creating files in quick succession,
use \code{"\%OSn"} as part of \code{format_stamp} to create precise stamps by truncating
seconds to \verb{0 <= n <= 6} decimal places, see \code{\link[=strftime]{strftime()}} for details.
}
\section{Side effects}{

The directory indicated by the returned file path is created if it does not
yet exist.
}

\examples{
# Use a temporary directory to not write in the user's directory
my_tempdir <- normalizePath(path = file.path(tempdir(), "subdir"),
                            winslash = "/", mustWork = FALSE)

(create_file_path(filename = "abc.txt", format_stamp = "",
                  dir = my_tempdir, add_date = TRUE))
(create_file_path(filename = "abc.txt", format_stamp = "\%d_\%m_\%Y",
                  dir = my_tempdir, add_date = TRUE))
(create_file_path(filename = "def.html", format_stamp = "",
                  dir = my_tempdir, add_date = FALSE))
(create_file_path(filename = "def.html", format_stamp = "\%d_\%m_\%Y",
                  dir = my_tempdir, add_date = FALSE))
(create_file_path(filename = "abc.txt", format_stamp = "",
                  dir = file.path(my_tempdir, "subdir"), add_date = TRUE))
(create_file_path(filename = "abc.txt", format_stamp = "\%d_\%m_\%Y",
                  dir = file.path(my_tempdir, "subdir"), add_date = TRUE))
(create_file_path(filename = "def.html", format_stamp = "",
                  dir = file.path(my_tempdir, "subdir"), add_date = FALSE))
(create_file_path(filename = "def.html", format_stamp = "\%d_\%m_\%Y",
                  dir = file.path(my_tempdir, "subdir"), add_date = FALSE))

# Cleaning up
unlink(x = my_tempdir, recursive = TRUE)
rm(my_tempdir)

}
\seealso{
\code{\link[=get_filename]{get_filename()}} to check if a file exists and is a unique match to a pattern,
\code{\link[=file.path]{file.path()}} to construct file paths in a platform-independent way,
\code{\link[=normalizePath]{normalizePath()}} to create absolute normalised paths,
\code{\link[=create_dir]{create_dir()}} to create a directory if it does not yet exist

Other functions to handle paths and directories:
\code{\link[=create_dir]{create_dir()}},
\code{\link[=create_tempdir]{create_tempdir()}},
\code{\link[=file_path_no_ext]{file_path_no_ext()}},
\code{\link[=get_filename]{get_filename()}},
\code{\link[=is_filename]{is_filename()}},
\code{\link[=is_path]{is_path()}}
}
\concept{functions to handle paths and directories}
