% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_path.R
\name{create_path}
\alias{create_path}
\title{Create a file path}
\usage{
create_path(
  filename,
  format_stamp = "\%Y_\%m_\%d_\%H_\%M_\%S",
  dir = file.path(".", "output"),
  add_date = TRUE
)
}
\arguments{
\item{filename}{A character string with the file name, including the file
extension like \code{.csv} or \code{.txt}.}

\item{format_stamp}{A character string indicating the \link[=strftime]{format} of
the stamp to be added in front of the file name. No stamp is added if
\code{format_stamp} is an empty string (i.e., \code{""}).}

\item{dir}{Non-empty character string containing the path to a directory that
should be created if it does not yet exist.}

\item{add_date}{\code{TRUE} or \code{FALSE}: create a subdirectory with the current
date in the \link[=strftime]{format} \code{YYYY_mm_dd}?}
}
\value{
The created file path, returned \link[=invisible]{invisibly}.
}
\description{
Create a file path, creating the indicated directory if it does not yet exist.
}
\details{
\code{filename} should contain a file extension: a dot followed by alphanumeric
characters until the end of the file name. An error occurs if no file
extension is present. Non-alphanumeric characters in \code{filename} before the
file extension are replaced by underscores, with a warning.

A warning is issued if the file indicated by the returned path already exists.
Use \code{"\%OSn"} as part of \code{format_stamp} to truncate seconds to \verb{0 <= n <= 6}
decimal places for precise stamps, see \code{\link[=strftime]{strftime()}} for details.

The directory for the returned path is \link[=create_dir]{created} if it does not
yet exist. By default it is a subdirectory with the current date in format
\code{YYYY_mm_dd} in directory \code{output} below the current working directory.

The returned path is \link[=normalizePath]{normalised} such that it still works
if the \link[=getwd]{working directory} changes. \code{"/"} instead of \code{"\\\\"} is used
as \link[=normalizePath]{winslash}, such that the returned path can be used in
Windows' file system.
}
\section{Side effects}{

The directory indicated by the returned file path is created if it does not
yet exist.
}

\section{Programming notes}{

File names that end in a dot are handled incorrectly by \code{create_path()}:
\code{\link[tools:fileutils]{tools::file_path_sans_ext()}} does not recognise the extension of file names
that end in a dot, whereas \code{\link[tools:fileutils]{tools::file_ext()}} does recognise such extension.
Therefore, \code{create_path("ff..txt")} produces the nonsense result ending in
\code{"ff__txt.txt"} instead of ending in \code{"ff_.txt"}.
}

\seealso{
\code{\link[=get_filename]{get_filename()}} to check if a file exists and is a unique match to a pattern;
\code{\link[=file.path]{file.path()}} to construct file paths in a platform-independent way;
\code{\link[=normalizePath]{normalizePath()}} to create absolute paths; \code{\link[=create_dir]{create_dir()}} used by this
function to create a directory if it does not yet exists.

Other functions to check paths and create directories: 
\code{\link{create_dir}()},
\code{\link{get_filename}()}
}
\concept{functions to check paths and create directories}
