% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tempdir.R
\name{create_tempdir}
\alias{create_tempdir}
\title{Create a temporary directory}
\usage{
create_tempdir(subdir = "subdir", error_if_exists = TRUE)
}
\arguments{
\item{subdir}{A \link[checkinput:is_character]{character string} with the
subdirectory to create inside \code{\link[=tempdir]{tempdir()}}.}

\item{error_if_exists}{\code{TRUE} or \code{FALSE}: throw an error if the temporary
directory already exists? The default \code{TRUE} prevents deleting files if the
created temporary directory is programmatically deleted later on.}
}
\value{
The \link[=normalizePath]{normalized} path to the created temporary directory,
returned \link[=invisible]{invisibly}.
}
\description{
Create a temporary directory that can safely be removed.
}
\details{
The temporary directory is created as a subdirectory inside \code{tempdir()} to
ensure that \link[=unlink]{removing} the created directory does not give problems:
removing the directory returned by \code{tempdir()} can give problems, for example
because \code{RStudio} also uses that directory.

It is possible to create subdirectories inside a not-yet existing directory
(e.g., to create \verb{<tempdir>/output/<date>} if \verb{<tempdir>/output} does not yet
exist.
}
\section{Side effects}{

The requested temporary directory is created if does not yet exist.
}

\examples{
tempdir()
# Create a directory inside the directory returned by 'tempdir()'
(create_tempdir())

# Error if the directory already exists and 'error_if_exists' is 'TRUE'
try(create_tempdir(error_if_exists = TRUE))
(create_tempdir(error_if_exists = FALSE))

}
\seealso{
\code{\link[=create_dir]{create_dir()}} to create (non-temporary) directories.

Other functions to handle paths and directories:
\code{\link[=create_dir]{create_dir()}},
\code{\link[=create_path]{create_path()}},
\code{\link[=file_path_sans_ext]{file_path_sans_ext()}},
\code{\link[=get_filename]{get_filename()}}
}
\concept{functions to handle paths and directories}
