% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tempdir.R
\name{create_tempdir}
\alias{create_tempdir}
\title{Create a temporary directory}
\usage{
create_tempdir(subdir = "subdir")
}
\arguments{
\item{subdir}{A \link[checkinput:is_character]{character string} with the name
of a \strong{not-yet} existing temporary subdirectory to be created inside
\code{\link[=tempdir]{tempdir()}}. \code{subdir} should meet the requirements for a valid
\link[=is_path]{path}.}
}
\value{
The \link[=normalizePath]{normalized} path to the created temporary directory,
returned \link[=invisible]{invisibly}.
}
\description{
Create a temporary directory that can safely be removed.
}
\details{
\code{subdir} is created inside \code{\link[=tempdir]{tempdir()}} and an error is thrown if it already
exists. This ensures that \link[=unlink]{removing} the created directory does not
remove files that are still needed by other processes (see \code{Usage} below).

It is possible to create subdirectories inside a not-yet existing directory
(e.g., to create \verb{<tempdir>/output/outputsub} if \verb{<tempdir>/output} does not
yet exist.
}
\section{Side effects}{

The requested temporary directory is created if does not yet exist. An error
is thrown if the directory already exists or creating the directory fails.
}

\section{Usage in practice}{

Examples and tests should \strong{not} write to the \link[=getwd]{working directory}
but to a temporary directory that is cleaned up afterwards. Although
\code{\link[=tempdir]{tempdir()}} points to a temporary directory, that directory should \strong{not} be
removed because \href{https://posit.co/products/open-source/rstudio}{RStudio} also
uses it. Instead, store the paths to temporary files written to \code{\link[=tempdir]{tempdir()}}
and \link{unlink} those paths when cleaning up, or create a temporary
subdirectory in \code{tempdir()} which can be completely be removed when cleaning
up. \code{use_tempdir()} follows the latter approach.
}

\examples{
tempdir()
# Create a directory inside the directory returned by 'tempdir()'
(tempdir_std <- create_tempdir(subdir = "examplesubtempdir"))

# Error if the directory already exists
try(create_tempdir(subdir = "examplesubtempdir"))

# It is possible to create recursive directories
(tempdir_recursive <- create_tempdir(subdir = file.path("abc", "def")))

# Clean up
unlink(c(tempdir_std, dirname(tempdir_recursive)), recursive = TRUE)
rm(tempdir_recursive, tempdir_std)

}
\seealso{
\code{\link[=create_dir]{create_dir()}} to create (non-temporary) directories.

Other functions to handle paths and directories:
\code{\link[=create_dir]{create_dir()}},
\code{\link[=create_file_path]{create_file_path()}},
\code{\link[=file_path_no_ext]{file_path_no_ext()}},
\code{\link[=get_file_path]{get_file_path()}},
\code{\link[=is_path]{is_path()}}
}
\concept{functions to handle paths and directories}
