% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tempdir.R
\name{create_tempdir}
\alias{create_tempdir}
\title{Create a temporary directory}
\usage{
create_tempdir(subdir = "subdir")
}
\arguments{
\item{subdir}{A \link[checkinput:is_character]{character string} with the
subdirectory to create inside \code{\link[=tempdir]{tempdir()}}. This subdirectory should not yet
exist, see \code{Details}.}
}
\value{
The \link[=normalizePath]{normalized} path to the created temporary directory,
returned \link[=invisible]{invisibly}.
}
\description{
Create a temporary directory that can safely be removed.
}
\details{
The temporary directory is created as subdirectory \code{subdir} inside
\code{tempdir()} and an error is thrown if it already exists. This ensures that
programmatically \link[=unlink]{removing} the created directory later on does not
remove files that are still needed by other processes (which would happen
when removing the directory returned by \code{tempdir()} because \code{RStudio} also
uses that directory).

It is possible to create subdirectories inside a not-yet existing directory
(e.g., to create \verb{<tempdir>/output/<date>} if \verb{<tempdir>/output} does not yet
exist.
}
\section{Side effects}{

The requested temporary directory is created if does not yet exist. An error
is thrown if the directory already exists or creating the directory failed.
}

\examples{
tempdir()
# Create a directory inside the directory returned by 'tempdir()'
(create_tempdir(subdir = "subdir"))

# Error if the directory already exists
try(create_tempdir(subdir = "subdir"))

# It is possible to create recursive directories
(create_tempdir(subdir = "abc/def"))

}
\seealso{
\code{\link[=create_dir]{create_dir()}} to create (non-temporary) directories and the notes in its
test-file \verb{progutils\\inst\\tinytest\\test_create_dir.R}.

Other functions to handle paths and directories:
\code{\link[=create_dir]{create_dir()}},
\code{\link[=create_path]{create_path()}},
\code{\link[=file_path_sans_ext]{file_path_sans_ext()}},
\code{\link[=get_filename]{get_filename()}}
}
\concept{functions to handle paths and directories}
