% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_path_no_ext.R
\name{file_path_no_ext}
\alias{file_path_no_ext}
\alias{file_path_ext}
\title{Extract filenames or extensions}
\usage{
file_path_no_ext(x, compression = FALSE)

file_path_ext(x, compression = FALSE)
}
\arguments{
\item{x}{character vector giving file paths.}

\item{compression}{logical: should compression extension \file{.gz},
    \file{.bz2} or \file{.xz} be removed first?}
}
\description{
Distinguish filenames and extensions, correctly handling filenames that end
in a dot.
}
\details{
\code{file_path_no_ext()} returns the file paths without extensions and leading
dot; \code{file_path_ext()} returns the file extensions without compression
extension and leading dot. Only purely alphanumeric extensions are recognized.

\code{file_path_no_ext()} replaces \code{\link[tools:file_path_sans_ext]{tools::file_path_sans_ext()}} because prior to
\R 4.6.0 the latter did \strong{not} recognise the extension of file names ending
in a dot, whereas \code{\link[tools:file_ext]{tools::file_ext()}} \strong{did} recognise such extensions. This
discrepancy led to \code{filename} \strong{not} being re-created by
\code{paste0(tools::file_path_sans_ext(filename), ".", tools::file_ext(filename))}
for filenames like \code{"ab..txt"}, instead producing the nonsense result
\code{"ab..txt.txt"}, see the \code{Examples}.

\code{file_path_ext()} replaces \code{\link[tools:file_ext]{tools::file_ext()}} because the latter does not
have argument \code{compression}.
}
\examples{
filename <- "ab..txt"
# Should be "ab." but was "ab..txt" prior to R 4.6.0.
tools::file_path_sans_ext(filename)
tools::file_ext(filename) # "txt"
# The next line produced the nonsense-result "ab..txt.txt" prior to R 4.6.0.
paste0(tools::file_path_sans_ext(filename), ".", tools::file_ext(filename))

# The updated version recreates filename 'ab..txt':
paste0(file_path_no_ext(filename), ".", file_path_ext(filename))

}
\seealso{
Other functions to handle paths and directories:
\code{\link[=create_dir]{create_dir()}},
\code{\link[=create_file_path]{create_file_path()}},
\code{\link[=create_tempdir]{create_tempdir()}},
\code{\link[=get_filename]{get_filename()}},
\code{\link[=is_filename]{is_filename()}},
\code{\link[=is_path]{is_path()}}
}
\concept{functions to handle paths and directories}
