% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_path_no_ext.R
\name{file_path_no_ext}
\alias{file_path_no_ext}
\title{Remove extension from file paths}
\usage{
file_path_no_ext(x, compression = FALSE)
}
\arguments{
\item{x}{character vector giving file paths.}

\item{compression}{logical: should compression extension \file{.gz},
    \file{.bz2} or \file{.xz} be removed first?}
}
\description{
A drop-in replacement for \code{\link[tools:file_path_sans_ext]{tools::file_path_sans_ext()}}, see \code{Details}.
Returns the file path without the extensions (and the leading dot). Only
purely alphanumeric extensions are recognized.
}
\details{
\code{\link[tools:file_path_sans_ext]{tools::file_path_sans_ext()}} did \strong{not} recognise the extension of file
names that end in a dot prior to \R 4.6.0, whereas \code{\link[tools:file_ext]{tools::file_ext()}}
\strong{did} recognise such extensions. This discrepancy led to \code{filename} not
being re-created by
\code{paste0(tools::file_path_sans_ext(filename), ".", tools::file_ext(filename))}
if \code{filename} ended in a dot.

Using \code{\link[=file_path_no_ext]{file_path_no_ext()}} ensures that \code{filename} is recreated by
\code{paste0(progutils::file_path_no_ext(filename), ".", tools::file_ext(filename))},
such that, e.g., \link{create_path}\code{("ab..txt")} produces the correct result
ending in \code{"ab..txt"} instead of the nonsense result ending in \code{"ab..txt.txt"},
see the \code{Examples}.
}
\examples{
filename <- "ab..txt"
# Should be "ab." but was "ab..txt" prior to R 4.6.0.
tools::file_path_sans_ext(filename)
tools::file_ext(filename) # "txt"
# So the next line produces the nonsense-result "ab..txt.txt"
paste0(tools::file_path_sans_ext(filename), ".", tools::file_ext(filename))

# The updated version recreates filename 'ab..txt':
paste0(progutils::file_path_no_ext(filename), ".", tools::file_ext(filename))

}
\seealso{
Other functions to handle paths and directories:
\code{\link[=create_dir]{create_dir()}},
\code{\link[=create_path]{create_path()}},
\code{\link[=create_tempdir]{create_tempdir()}},
\code{\link[=get_filename]{get_filename()}},
\code{\link[=is_filename]{is_filename()}},
\code{\link[=is_path]{is_path()}}

Other functions to modify character vectors:
\code{\link[=as.numeric_safe]{as.numeric_safe()}},
\code{\link[=reorder_cols]{reorder_cols()}},
\code{\link[=replace_vals]{replace_vals()}},
\code{\link[=unpaste_unquote]{unpaste_unquote()}},
\code{\link[=vect_to_char]{vect_to_char()}},
\code{\link[=wrap_text]{wrap_text()}}
}
\concept{functions to handle paths and directories}
\concept{functions to modify character vectors}
