% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_path_sans_ext.R
\name{file_path_sans_ext}
\alias{file_path_sans_ext}
\title{Remove extension from file paths}
\usage{
file_path_sans_ext(x, compression = FALSE)
}
\arguments{
\item{x}{character vector giving file paths.}

\item{compression}{logical: should compression extension \file{.gz},
    \file{.bz2} or \file{.xz} be removed first?}
}
\description{
A drop-in replacement for tools::file_path_sans_ext, see \code{Details}. Returns
the file path without the extensions (and the leading dot). Only purely
alphanumeric extensions are recognized.
}
\details{
\code{\link[tools:fileutils]{tools::file_path_sans_ext()}} does \emph{not} recognise the extension of file names
that end in a dot, whereas \code{\link[tools:fileutils]{tools::file_ext()}} \emph{does} recognise such
extensions.

Using \code{\link[=file_path_sans_ext]{file_path_sans_ext()}} from \code{progutils} ensures that
\code{filename} is recreated by
\code{paste0(progutils::file_path_sans_ext(filename), ".", tools::file_ext(filename))},
such that, e.g., \link{create_path}\code{("ab..txt")} produces the correct result
ending in \code{"ab..txt"} instead of the nonsense result ending in \code{"ab..txt.txt"},
see the \code{Examples}.
}
\examples{
filename <- "ab..txt"
tools::file_path_sans_ext(filename) # "ab..txt"
tools::file_ext(filename) # "txt"
# So the next line produces the nonsense-result "ab..txt.txt"
paste0(tools::file_path_sans_ext(filename), ".", tools::file_ext(filename))

# The updated version recreates filename 'ab..txt':
paste0(progutils::file_path_sans_ext(filename), ".", tools::file_ext(filename))

}
