% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_path.R
\name{get_file_path}
\alias{get_file_path}
\title{Check that only a single matching file is present}
\usage{
get_file_path(dir = ".", pattern, ignore_case = TRUE, quietly = FALSE)
}
\arguments{
\item{dir}{Character string with the \link[checkinput:is_path]{path} to a
directory.}

\item{pattern}{Character string containing a \link[base:regex]{regular expression}
used to select names of files that are present in \code{dir}.}

\item{ignore_case}{\code{TRUE} or \code{FALSE}: use case-insensitive pattern matching?}

\item{quietly}{\code{TRUE} or \code{FALSE}: suppress the message with the found file
name?}
}
\value{
A character string with the \link[fs:path_abs]{absolute normalized} path to the
file with a name matching \code{pattern} if there is exactly one such file in
directory \code{dir}. Otherwise an error is thrown. Use \code{\link[=basename]{basename()}} on the result
to obtain the filename itself.
}
\description{
Check that only one file in a directory has a name matching \code{pattern}, for
example before attempting to \link[utils:read.table]{read a file}.
}
\details{
The default \code{dir} (\code{"."}) indicates the \link[=getwd]{working directory}.

If \code{ignore_case} is \code{FALSE} and no case-sensitive match is found, the error
message indicates if any case-insensitive match is present.

In contrast to the default of \code{\link[=list.files]{list.files()}}, \code{get_file_path()} also finds
'hidden' files, i.e., files with names that start with a dot.

Paths will be \link[fs:path_abs]{normalized} to ensure they still work if the
\link[=getwd]{working directory} changes.
}
\examples{
# Create files in a temporary directory so we know what is present.
my_tempfiles <- tempfile(pattern = c("some_filename", "another_filename"),
                         fileext = ".txt")
# Create the files
file.create(my_tempfiles)

get_file_path(dir = tempdir(), pattern = "some_file")

# The same file is found if case-insensitive matching is used:
get_file_path(dir = tempdir(), pattern = "SOME_FILE", ignore_case = TRUE)

# Error reporting the presence of a case-insensitive match.
try(get_file_path(dir = tempdir(), pattern = "SOME_FILE", ignore_case = FALSE))

# 'pattern' is interpreted as a regular expression
get_file_path(dir = tempdir(), pattern = "^.+er_file")

# Error reporting no match found.
try(get_file_path(dir = tempdir(), pattern = "missing_filename_abcde",
                 ignore_case = TRUE))
try(get_file_path(dir = tempdir(), pattern = "missing_filename_abcde",
                 ignore_case = FALSE))

# Error if multiple matches are present.
try(get_file_path(dir = tempdir(), pattern = "_filename"))

# Clean up
unlink(x = my_tempfiles)
rm(my_tempfiles)

}
\seealso{
\code{\link[checkinput:is_path]{checkinput::is_path()}} to check if a path is valid, and the 'Note on paths'
in its documentation;
\code{\link[=create_dir]{create_dir()}} to create a directory if does not yet exist;
\code{\link[=file.exists]{file.exists()}} and \code{\link[=list.files]{list.files()}} to check for existence of files without
checking they are a unique match to a pattern;
\code{\link[=file.info]{file.info()}} and \code{\link[=file.access]{file.access()}} to extract information about files or
directories;
\code{\link[fs:path]{fs::path()}} to construct file paths in a platform-independent way;
\code{\link[fs:path_abs]{fs::path_abs()}} to create absolute paths.

Other functions to handle paths and directories:
\code{\link[=create_dir]{create_dir()}},
\code{\link[=create_file_path]{create_file_path()}},
\code{\link[=create_tempdir]{create_tempdir()}}

Other functions to check equality:
\code{\link[=are_equal]{are_equal()}},
\code{\link[=check_case]{check_case()}},
\code{\link[=not_in]{not_in()}},
\code{\link[=replace_vals]{replace_vals()}}
}
\concept{functions to check equality}
\concept{functions to handle paths and directories}
