% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paths.R
\name{get_paths}
\alias{get_paths}
\title{Get paths to directories}
\usage{
get_paths(paths = .libPaths())
}
\arguments{
\item{paths}{The paths to use.}
}
\value{
A character vector containing the same elements as \code{\link[=.libPaths]{.libPaths()}}, possibly in
a different order.
}
\description{
Get the paths \R knows about, with paths containing the currently used \R
version number given before other paths.
}
\details{
This is a wrapper around \code{\link[=.libPaths]{.libPaths()}} to return paths containing the
currently used \R \link[=getRversion]{version number} in the form of directory
\code{R-x.y.z} or directory \code{x.y}, with a warning if none such path is present,
before other paths.

A warning is issued if the returned vector is of length zero or only contains
empty paths (\code{""}).
}
\examples{
.libPaths()
get_paths() # The same elements as above, possibly in a different order
getRversion() # The currently used R version

}
\seealso{
Other functions to check paths and create directories: 
\code{\link{create_dir}()},
\code{\link{get_filename}()}
}
\concept{functions to check paths and create directories}
