% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_filename.R
\name{is_filename}
\alias{is_filename}
\title{Check that \code{x} is a valid filename}
\usage{
is_filename(filename)
}
\arguments{
\item{filename}{\link[checkinput:is_character]{character string} with the
filename including the file extension.}
}
\value{
\code{TRUE}: an error occurs if \code{filename} is not a valid filename.
}
\description{
Check that \code{x} is a filename that includes an extension and is likely to be
valid.
}
\details{
\code{is_filename()} puts some restrictions on filenames so it can be used to
check for valid filenames before creating a file:
\itemize{
\item \code{filename} should include a file extension. The compression extensions
\code{".gz"}, \code{".bz2"} or \code{".xz"} are \strong{not} considered file extensions but are
allowed to be present.
\item \code{filename} should \strong{not} contain characters \verb{"}, \code{*}, \code{/}, \code{:}, \verb{?}, \verb{\\},
\code{|}, \code{<}, \code{>}, nor any of the control characters (\code{ASCII} octal codes 000
through 037 and 177, see \code{help("regex")}).
\item \code{filename} should \strong{not} start with a space.
\item The part of \code{filename} before the extension should \strong{not} end with a space
nor with a period.
}

These restrictions consider characters that would lead to an error in Windows
because they are not allowed; characters that  would lead to a mismatch
between the created directory and the returned path because they are silently
removed on Windows; and words that are reserved names in Windows.

In contrast to functions from \code{checkinput}, \code{is_filename} will produce an
error if \code{filename} is not a valid filename.
}
\section{Programming notes}{

Although not enforced by \code{is_filename()}, it is good practice to also avoid
the characters \code{+}, \verb{,}, \verb{;}, \code{=}, \code{[}, \verb{]}, \code{!}, \code{$}, \verb{#}, \code{@}, and possibly
\verb{\{}, \verb{\}}, \code{(}, \verb{)}, \verb{'}, \verb{\%}, \code{&}, \if{html}{\out{<backtick>}}, \code{^}, \code{~}.

Ways to make \code{is_filename()} even stricter:
\itemize{
\item do not allow filenames to start with a hyphen
\item do not allow filenames to end with a hyphen
\item case-insensitive matching to \code{filename} to determine if it exists?
\code{filename} should \strong{not} point to a directory (see \code{utils::file_test()},
\code{\link[=get_filename]{get_filename()}}, \code{\link[=create_tempdir]{create_tempdir()}}, \code{\link[=create_file_path]{create_file_path()}}).
\item Impose a limit on the length:
https://blog.r-project.org/2023/03/07/path-length-limit-on-windows/
\item See also functions \code{create_file_path()}, \code{make_filename()} and
\code{fs::path_real()}.
}
}

\section{References}{

\itemize{
\item https://learn.microsoft.com/en-us/windows/win32/fileio/naming-a-file
\item https://en.wikipedia.org/wiki/Comparison_of_file_systems#Limits
}
}

\examples{
is_filename("abcd.txt")
is_filename("abcd.txt.gz")

try(is_filename("abcd"))
try(is_filename("abcd.gz"))
try(is_filename("ab|cd.txt"))

}
\seealso{
On file existence and permissions, see \code{utils::file_test()} and references
there and \code{fs::path_real()}.

Other functions to handle paths and directories:
\code{\link[=create_dir]{create_dir()}},
\code{\link[=create_file_path]{create_file_path()}},
\code{\link[=create_tempdir]{create_tempdir()}},
\code{\link[=file_path_no_ext]{file_path_no_ext()}},
\code{\link[=get_filename]{get_filename()}},
\code{\link[=is_path]{is_path()}}
}
\concept{functions to handle paths and directories}
