% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_path.R
\name{is_path}
\alias{is_path}
\title{Check that \code{x} is a valid path}
\usage{
is_path(path, as_file = FALSE)
}
\arguments{
\item{path}{\link[checkinput:is_character]{character string} with the path.}

\item{as_file}{\code{TRUE} or \code{FALSE}: is \code{path} intended to point to a file
instead of to a folder?}
}
\value{
\code{TRUE}: an error occurs if \code{path} is not a valid path.
}
\description{
Check that \code{x} is a path that s likely to be a valid path.
}
\details{
\code{is_path()} puts some restrictions on paths so it can be used to check for
valid paths before creating a directory:
\itemize{
\item \code{path} should \strong{not} contain the characters \verb{"}, \code{*}, \verb{?}, \code{|}, \code{<}, \code{>},
nor any of the control characters (\code{ASCII} octal codes 000 through 037 and
177, see \code{help("regex")}).
\item \code{path} components should \strong{not} end with a slash, backslash, space or dot
(\code{"."} and \code{".."} are allowed as first component to indicate the working
directory and the parent directory, respectively).
\item \code{path} components should \strong{not} be \code{CON}, \code{PRN}, \code{AUX}, \code{NUL},
\verb{COM<non-zero digit>}, \verb{LPT<non-zero digit>}, case-insensitive variants of
these names, and these names followed by an extension.
\item \code{path} should not point to \code{tempdir()}: a temporary subdirectory should be
used instead.
}

These restrictions consider characters that would lead to an error in Windows
because they are not allowed; characters that  would lead to a mismatch
between the created directory and the returned path because they are silently
removed in Windows; and words that are reserved names in Windows.

In contrast to functions from \code{checkinput}, \code{is_path} will produce an error
if \code{path} is not a valid path.
}
\section{References}{

\itemize{
\item https://learn.microsoft.com/en-us/windows/win32/fileio/naming-a-file
\item https://en.wikipedia.org/wiki/Comparison_of_file_systems#Limits
}
}

\examples{
is_path(getwd())
try(is_path(file.path(getwd(), "ab|cd")))

}
\seealso{
\code{\link[=create_path]{create_path()}} to create a path (with references there about file paths),
\code{\link[=create_dir]{create_dir()}} to create a directory if it does not yet exist,
\code{\link[=get_filename]{get_filename()}} to check if a file exists and is a unique match to a pattern

\code{fs::path_sanitize()} to \emph{remove} invalid characters from potential paths.

Other functions to handle paths and directories:
\code{\link[=create_dir]{create_dir()}},
\code{\link[=create_path]{create_path()}},
\code{\link[=create_tempdir]{create_tempdir()}},
\code{\link[=file_path_no_ext]{file_path_no_ext()}},
\code{\link[=get_filename]{get_filename()}},
\code{\link[=is_filename]{is_filename()}}
}
\concept{functions to handle paths and directories}
