% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{not_in}
\alias{not_in}
\title{Check that values are absent}
\usage{
not_in(x, table, value = TRUE)
}
\arguments{
\item{x}{Vector or factor with values to test absence from \code{table}. \code{x}
should not be of \link[=typeof]{type} \code{double}.}

\item{table}{Vector or factor in which to test for absence of \code{x}. \code{table}
should not be of \code{type} \code{double}.}

\item{value}{\code{TRUE} or \code{FALSE}: should a vector with values be returned
instead of a boolean vector?}
}
\value{
If \code{value} is \code{TRUE}: the values in \code{x} that are absent from \code{table}, with a
zero-length object of the same type as \code{x}, e.g., \code{character(0)} or
\code{logical(0)} if none of the values in \code{x} are absent from \code{table} (i.e., all
are present in \code{table}). If \code{value} is \code{FALSE}: a boolean vector indicating
for each element in \code{x} if it is absent from \code{table}.
}
\description{
Check that values from one vector are absent from another vector
}
\details{
Duplicates in \code{x} are kept, in contrast to \code{\link[=setdiff]{setdiff()}}, see the \code{Examples}.

\link[=factor]{Factor-input} to \code{x} is converted to character, to prevent
returning a factor with all values of \code{x} as \link{levels}.

Zero-length input behaves slightly different from other values: \code{not_in()}
returns \code{logical(0)} for zero-length input to \code{x} if that is absent from
\code{table} if \code{value} is \code{FALSE}. If \code{value} is \code{TRUE}, the behaviour is normal:
returning \code{x}.

\link{NA} is allowed in \code{x} and \code{table} and behaves the same as other values: the
returned \code{NA} (if \code{value} is \code{TRUE}) and the returned zero-length value (if
\code{value} is \code{FALSE}) have the same type as the \code{NA} in \code{x} if \code{NA} is
absent from \code{table}. \code{NA}s of different types in \code{x} and \code{table} match each
other.

\code{not_in()} does \strong{not} allow input of \link[=typeof]{type} \code{double} because
matching such input should allow for small numerical errors by using a
tolerance, for example, as the error message indicates, using \code{\link[=are_equal]{are_equal()}}.

Apart from \strong{not} allowing numeric input, \code{not_in(x, table, value = FALSE)}
is equivalent to \code{x \%notin\% table}, where \verb{\%notin\%} is a function in base \R
since version \verb{4.6.0}.

Names are \strong{not} considered when matching but are retained in the output,
similar to \code{\%in\%}.
}
\examples{
x <- letters[1:4]
table <- letters[3:6]
not_in(x, table) # c("a", "b")
not_in(as.factor(x), as.factor(table)) # c("a", "b")
# c(TRUE, TRUE, FALSE, FALSE), same as !(x \%in\% table):
not_in(x, table, value = FALSE)

x_dupl <- c(x, letters[c(2, 4:6, 5)])
table_dupl <- letters[c(3:8, 5:7)]
not_in(x_dupl, table_dupl) # c("a", "b", "b")
setdiff(x_dupl, table_dupl) # c("a", "b")
not_in(x_dupl, table_dupl, value = FALSE)
# c(TRUE, TRUE, FALSE, FALSE, TRUE, FALSE, FALSE, FALSE, FALSE)

# Names are not considered when matching but are retained in the output
not_in(c(x = "c", y = "b", z = "a"), c(a = "a", b = "b"))

}
\seealso{
\code{\link[=setdiff]{setdiff()}} for a similar function which removes duplicates;
\code{\link[=are_equal]{are_equal()}} to match numeric input using a tolerance;
\code{\link[=identical]{identical()}}, \code{\link[=not_in]{not_in()}} and \code{\link[=match]{match()}} (containing \code{\%in\%} and, from \R
\verb{4.6.0} onwards, \code{'\%notin\%'}) to check for exact (in)equality, with
\code{\link{Comparison}} to do so using binary operators

Other functions to check equality:
\code{\link[=are_equal]{are_equal()}},
\code{\link[=check_case]{check_case()}},
\code{\link[=get_file_path]{get_file_path()}},
\code{\link[=replace_nonalnum]{replace_nonalnum()}},
\code{\link[=replace_vals]{replace_vals()}},
\code{\link[=round_factor]{round_factor()}},
\code{\link[=signif_custom]{signif_custom()}}
}
\concept{functions to check equality}
