% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{not_in}
\alias{not_in}
\title{Check if values from a vector are not present in another vector}
\usage{
not_in(x, y, value = TRUE)
}
\arguments{
\item{x}{Vector with values to look for absence from \code{y}.}

\item{y}{Vector in which to look for absence of values \code{x}.}

\item{value}{\code{TRUE} or \code{FALSE}: should a vector with values be returned
instead of a boolean vector?}
}
\value{
If \code{value} is \code{TRUE}: the values in \code{x} that are absent from \code{y} or, if none
of the values in \code{x} are absent from \code{y} (i.e., all are present in \code{y}), a
zero-length object of the same type as \code{x}, e.g., \code{character(0)} or
\code{logical(0)}. If \code{value} is \code{FALSE}: a boolean vector indicating for each
element in \code{x} if it is absent from \code{y}, i.e., \code{!(x \%in\% y)}.
}
\description{
Check if values from a vector are not present in another vector
}
\details{
Duplicates in \code{x} are kept, in contrast to \code{\link[=setdiff]{setdiff()}}, see the \code{Examples}.

\link[=factor]{Factor-input} to \code{x} is converted to character, to prevent
returning a factor with all values of \code{x} as \link{levels}.

\link{NA}s of different types match each other.
}
\section{Programming note}{

\code{not_in(x, y, value = FALSE)} gives the same output as function \verb{\%notin\%}
that is present in \R\verb{-devel base} since 05/12/2025 with definition
\code{match(x, table, nomatch = 0L) == 0L}.
}

\examples{
x <- letters[1:4]
y <- letters[3:6]
not_in(x, y) # c("a", "b")
not_in(as.factor(x), as.factor(y)) # c("a", "b")
not_in(x, y, value = FALSE) # c(TRUE, TRUE, FALSE, FALSE), same as !(x \%in\% y)

x_dupl <- c(x, letters[c(2, 4:6, 5)])
y_dupl <- letters[c(3:8, 5:7)]
not_in(x_dupl, y_dupl) # c("a", "b", "b")
setdiff(x_dupl, y_dupl) # c("a", "b")
not_in(x_dupl, y_dupl, value = FALSE)
# c(TRUE, TRUE, FALSE, FALSE, TRUE, FALSE, FALSE, FALSE, FALSE)

}
\seealso{
\code{\link[=match]{match()}} \code{\link[=setdiff]{setdiff()}} \verb{\%notin\%} (from \R 4.6.0 onwards)
}
