% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_quoted.R
\name{paste_quoted}
\alias{paste_quoted}
\title{Concatenate x to a string with quoted elements.}
\usage{
paste_quoted(x)
}
\arguments{
\item{x}{Vector to be converted to a character string.}
}
\value{
A character string consisting of the elements of \code{x} surrounded by single
quotes, separated by commas. \code{NULL} is returned as \code{"'NULL'"}, other
zero-length objects are returned as \code{"'<class>(0)'"}, e.g., \code{"'logical(0)'"}.
}
\description{
Concatenate a vector to a string with quoted elements, returning \code{NULL} as
\code{"'NULL'"} and other zero-length objects as \code{"'<class>(0)'"}, e.g.,
\code{"'logical(0)'"}.
}
\note{
An error occurs if multiple arguments are provided because then \code{x} probably
was accidentally not \link[=c]{combined}. For example, the call
\code{paste_quoted("a", "b")} will return the error \verb{unused argument ("b")}. The
probably intended call is \code{paste_quoted(c("a", "b"))}, returning \code{"'a', 'b'"}.
}
\examples{
paste_quoted(c(3, 4)) # "'3', '4'"
paste_quoted(NULL) # "'NULL'"
paste_quoted(c(a = 3, b = 4)) # "'3', '4'" # Warns about dropping names.

}
\seealso{
\code{\link[=sQuote]{sQuote()}}; \code{\link[=toString]{toString()}} which can be used instead of
\code{paste(x, collapse = ", ")}; \code{\link[=paste0]{paste0()}}.

Other functions to modify character vectors: 
\code{\link{as.numeric_safe}()},
\code{\link{vect_to_char}()},
\code{\link{wrap_text}()}
}
\concept{functions to modify character vectors}
