% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_quoted.R
\name{paste_quoted}
\alias{paste_quoted}
\title{Concatenate text to a single string of quoted elements.}
\usage{
paste_quoted(x)
}
\arguments{
\item{x}{Vector to be converted to a single character string.}
}
\value{
A character string consisting of the elements of \code{x} surrounded by
quotes (i.e., \verb{'}), separated by commas (i.e., \verb{,}).
}
\description{
Concatenate text to a single string of quoted elements, printing \code{NULL} and
\code{character(0)} as \code{"'NULL'"}.
}
\note{
\code{NULL} and \code{character(0)} are printed as \code{"'NULL'"}. Other zero-length
objects are printed as \code{"''"}.

An error occurs if multiple arguments are provided. Then \code{x} was probably
accidentally not concatenated. For example, the call \code{paste_quoted("a", "b")}
will return the error \verb{unused argument ("b")}. The probably intended call is
\code{paste_quoted(c("a", "b"))}.
}
\section{To do}{

See also \code{\link[=dQuote]{dQuote()}}, \code{\link[=sQuote]{sQuote()}} and \code{\link[=Quotes]{Quotes()}}
}

\section{Wishlist}{
 Implement preserving names, e.g., using
\code{paste(names(x), x, sep = ": ", collapse = ", ")} (used in \code{\link[=vect_to_char]{vect_to_char()}})
or \code{paste0("'", paste(names(x), x, sep = ": ", collapse = "', '"), "'")}. Add
argument 'use_names = c("numeric", "all", "none") to use that?
}

\examples{
paste_quoted(c(3, 4)) # "'3', '4'"
paste_quoted(NULL) # "'NULL'"
paste_quoted(c(a = 3, b = 4)) # "'3', '4'" # Warns about dropping names.

}
\seealso{
\code{\link[=paste0]{paste0()}} \code{\link[=toString]{toString()}} which can be used instead of
\code{paste(x, collapse = ", ")}

Other functions to modify character vectors: 
\code{\link{as.numeric_safe}()},
\code{\link{vect_to_char}()},
\code{\link{wrap_text}()}
}
\concept{functions to modify character vectors}
