% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_quoted.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{paste_quoted}
\title{Objects exported from other packages}
\seealso{
Other functions to convert types:
\code{\link[=as.numeric_safe]{as.numeric_safe()}},
\code{\link[=reorder_levels]{reorder_levels()}},
\code{\link[=vect_to_char]{vect_to_char()}}

Other functions to modify character vectors:
\code{\link[=as.numeric_safe]{as.numeric_safe()}},
\code{\link[=replace_vals]{replace_vals()}},
\code{\link[=signal_text]{signal_text()}},
\code{\link[=unpaste_unquote]{unpaste_unquote()}},
\code{\link[=vect_to_char]{vect_to_char()}},
\code{\link[=wrap_text]{wrap_text()}}

Other functions to modify factors:
\code{\link[=as.numeric_safe]{as.numeric_safe()}},
\code{\link[=reorder_levels]{reorder_levels()}},
\code{\link[=replace_vals]{replace_vals()}},
\code{\link[=round_factor]{round_factor()}},
\code{\link[=vect_to_char]{vect_to_char()}}
}
\concept{functions to convert types}
\concept{functions to modify character vectors}
\concept{functions to modify factors}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{checkinput}{\code{\link[checkinput:paste_quoted]{paste_quoted()}}}
}}

