% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_quoted.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{paste_quoted}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{checkinput}{\code{\link[checkinput:paste_quoted]{paste_quoted()}}}
}}

