% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_levels.R
\name{reorder_levels}
\alias{reorder_levels}
\title{Reorder factor levels}
\usage{
reorder_levels(x, new_order, warn_drop_order = TRUE)
}
\arguments{
\item{x}{factor with levels to be reordered to \code{new_order}, or character
vector to be converted to a factor with levels ordered as \code{new_order}. Should
have length larger than zero.}

\item{new_order}{a \link[checkinput:all_characters]{character vector} with a
length larger than zero containing \link{unique} levels giving the desired order,
or \link{NULL} to sort levels on increasing numerical value.}

\item{warn_drop_order}{\code{TRUE} or \code{FALSE}: warn if values of \code{new_order} are
dropped because they are not present in \code{x}?}
}
\value{
\code{x} after converting it to a \link{factor} with levels reordered to
\code{new_order}, dropping values in \code{new_order} that are not present in \code{x}.
}
\description{
Reorder factor levels
}
\details{
Character input to \code{x} is silently converted to a \link{factor} before reordering
the factor levels.

Values of factor \code{x} that are not present in its levels are added to its
levels, and levels that are not present in its values are dropped, both with
a warning.

Levels of \code{x} that are missing from \code{new_order} are appended to \code{new_order},
with a warning. Values of \code{new_order} that are missing from levels of \code{x} are
dropped, with a warning if \code{warn_drop_order} is \code{TRUE}.

Sorting on on increasing numerical value only works if \code{x} can be suitably
converted by \code{\link[=as.numeric_safe]{as.numeric_safe()}}.
}
\section{Notes}{

Reordering levels of factor \code{f} by replacing levels through code like
\verb{levels(f) <- levels(f)[<some order>]} does \strong{not} work, see the last
\code{Example}.
}

\examples{
orig <- factor(letters[c(12:13, 13:11)], levels = letters[13:11])
orig
reorder_levels(x = orig, new_order = letters[11:13])

x_num <- factor(c(28, 3:2), levels = c(28, 3:2))
# Correctly sorts level 28 after 2 and 3
reorder_levels(x_num, new_order = NULL)

# Incorrectly sorts level "28" between "2" and "3"
sort(levels(x_num))

# Changing the levels directly does *not* work because it changes the values
levels(orig) <- letters[11:13]
orig

}
\seealso{
\code{\link[stats:relevel]{stats::relevel()}} to assign one reference level to a factor;
\code{\link[=order]{order()}}; \code{\link[=sort]{sort()}}

Other functions to convert types:
\code{\link[=as.numeric_safe]{as.numeric_safe()}},
\code{\link{reexports}},
\code{\link[=vect_to_char]{vect_to_char()}}

Other functions to modify factors:
\code{\link[=as.numeric_safe]{as.numeric_safe()}},
\code{\link{reexports}},
\code{\link[=replace_vals]{replace_vals()}},
\code{\link[=vect_to_char]{vect_to_char()}}

Other functions to modify sorting order:
\code{\link[=reorder_cols]{reorder_cols()}}
}
\concept{functions to convert types}
\concept{functions to modify factors}
\concept{functions to modify sorting order}
