% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_levels.R
\name{reorder_levels}
\alias{reorder_levels}
\title{Reorder factor levels}
\usage{
reorder_levels(x, new_order, warn_drop_order = TRUE)
}
\arguments{
\item{x}{Factor with levels to be reordered to \code{new_order}, or character
vector to be converted to a factor with levels ordered as \code{new_order}. Should
have length larger than zero.}

\item{new_order}{Unique character vector with a length larger than zero
indicating the new order of the factor levels.}

\item{warn_drop_order}{\code{TRUE} or \code{FALSE}: warn if values of \code{new_order} are
dropped because they are not present in \code{x}?}
}
\value{
\code{x} after converting it to a \link{factor} with its levels reordered to
\code{new_order}, dropping values in \code{new_order} not present in \code{x}.
}
\description{
Reorder factor levels
}
\details{
Character input to \code{x} is silently converted to a \link{factor} before reordering
the factor levels.

Values of factor \code{x} that are not present in its levels are added to its
levels, with a warning.

Levels of factor \code{x} that are not present in its values are dropped, with a
warning.

Levels of \code{x} that are missing from \code{new_order} are appended to \code{new_order},
with a warning.

Values in \code{new_order} that are missing from levels of \code{x} are dropped, with a
warning if \code{warn_drop_order} is \code{TRUE}.
}
\section{Notes}{

Reordering levels of factor \code{f} by replacing levels through code like
\verb{levels(f) <- levels(f)[<some order>]} does \emph{not} work, see the last \code{Example}.
}

\examples{
orig <- factor(letters[c(12:13, 13:11)], levels = letters[13:11])
orig
reorder_levels(x = orig, new_order = letters[11:13])

# Changing the levels directly does *not* work because it changes the values
levels(orig) <- letters[11:13]
orig

}
\seealso{
\link{reorder_cols} \code{\link[stats:relevel]{stats::relevel()}} to assign one reference level to a factor

Other functions to modify factors: 
\code{\link{as.numeric_safe}()},
\code{\link{replace_vals}()},
\code{\link{vect_to_char}()}
}
\concept{functions to modify factors}
