% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_nonalnum.R
\name{replace_nonalnum}
\alias{replace_nonalnum}
\title{Replace non-alphanumeric characters}
\usage{
replace_nonalnum(x, replacement = "_", keep_underscore = TRUE)
}
\arguments{
\item{x}{\link[checkinput:all_characters]{character vector} with the (possibly
empty) strings to remove non-alphanumeric characters from.}

\item{replacement}{\link[checkinput:is_character]{character string} used to
replace non-alphanumeric characters. It might be empty (i.e., \code{""}) to remove
instead of replace non-alphanumeric characters.}

\item{keep_underscore}{\code{TRUE} or \code{FALSE}: keep underscores? If \code{FALSE},
underscores are replaced by \code{replacement}, which only makes sense if a
non-default \code{replacement} is used such that it is not an underscore, see the
last \code{Example}.}
}
\value{
\code{x} with non-alphanumeric characters other than dots, and possibly
underscores, replaced by \code{replacement}.
}
\description{
Replace or remove non-alphanumeric characters, keeping dots and handling
underscores as indicated by \code{keep_underscores}.
}
\details{
This implementation uses the \link[=regex]{character class} \verb{[:alnum:]} that
depends on the current \link[=locales]{locale}, see the \verb{Programming notes} in
\code{\link[checkinput:all_names]{checkinput::all_names()}}.

Empty strings are allowed as input to \code{x} because they might also be returned
by \code{replace_nonalnum()} if the empty character string (\code{""}) is used for
\code{replacement}.
}
\examples{
replace_nonalnum("a+b.txt")
replace_nonalnum("a b.txt")
replace_nonalnum("ab12._")  # returned unchanged
replace_nonalnum(c("a+b.txt", "a b.txt", "ab12._"))

# Removing instead of replacing nonalphanumeric characters
replace_nonalnum("a+b.txt", replacement = "") # "ab.txt"

# Handling underscores
replace_nonalnum("a+bc_d.txt", replacement = "", keep_underscore = TRUE)  # "abc_d.txt"
replace_nonalnum("a+bc_d.txt", replacement = "", keep_underscore = FALSE) # "abcd.txt"
# Seems to keep the underscore because the default 'replacement' is an underscore
replace_nonalnum("a+bc_d.txt", keep_underscore = FALSE) # "a_bc_d.txt"

}
\seealso{
\code{\link[=trimws]{trimws()}} to remove leading and/or trailing whitespace;
\code{\link[checkinput:all_names]{checkinput::all_names()}} and \code{\link[checkinput:is_path]{checkinput::is_path()}} for other relevant
checks.

Other functions to check equality:
\code{\link[=are_equal]{are_equal()}},
\code{\link[=check_case]{check_case()}},
\code{\link[=get_file_path]{get_file_path()}},
\code{\link[=not_in]{not_in()}},
\code{\link[=replace_vals]{replace_vals()}},
\code{\link[=round_levels]{round_levels()}},
\code{\link[=signif_custom]{signif_custom()}}
}
\concept{functions to check equality}
