% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_vals.R
\name{replace_vals}
\alias{replace_vals}
\title{Replace character values or factor levels.}
\usage{
replace_vals(
  x,
  old,
  new,
  ignore_case = FALSE,
  allow_multiple = TRUE,
  warn_absent = TRUE,
  warn_case_new = c("always", "ignore_case", "case_sensitive", "never"),
  warn_case_old = c("case_sensitive", "never", "always", "ignore_case"),
  quiet = FALSE
)
}
\arguments{
\item{x}{Character vector with values to be replaced, or factor with levels
to be replaced.}

\item{old}{Character vector with values to be replaced. Should be unique and
not contain the value in \code{new}.}

\item{new}{Character string of length one with the new value.}

\item{ignore_case}{\code{TRUE} or \code{FALSE}: ignore case when looking for \code{old}?}

\item{allow_multiple}{\code{TRUE} or \code{FALSE}: allow multiple values of \code{old} to
match \code{x}? If \code{FALSE}, an error is thrown if multiple values of \code{old} match \code{x}.}

\item{warn_absent}{\code{TRUE} or \code{FALSE}: warn if \code{old} nor \code{new} are found in
\code{x}? If \code{new} is found, it is silently assumed replacement is not necessary
anymore.}

\item{warn_case_new, warn_case_old}{Character strings indicating when a warning
should be issued if values in \code{x} are a case-insensitive match but not a
case-sensitive match to \code{new} or \code{old}, respectively. Possible values are
\code{"always"} to always warn, \code{"ignore_case"} to only warn if \code{ignore_case} is
\code{TRUE}, \code{"case_sensitive"} to only warn if \code{ignore_case} is \code{FALSE} and
\code{"never"} to never warn.}

\item{quiet}{\code{TRUE} or \code{FALSE}: suppress printing a message with the values
that have been replaced?}
}
\value{
\code{x} with the requested replacements.
}
\description{
Replace character values or factor levels.
}
\details{
\code{ignore_case} does not affect \code{new}, e.g.,
\code{replace_vals(x = c("a", "B"), old = "a", new = "b", ignore_case = TRUE)}
returns \code{c("b", "B")}, \emph{not} \code{c("b", "b")}. Could add an argument to add
\code{tolower(new)} to \code{old} if \code{ignore_case} is \code{TRUE}?

If \code{allow_multiple} is \code{FALSE}, an error is thrown if multiple values in \code{old}
match \code{x}. This is useful if \code{old} should only be replaced with \code{new} if a
single value in \code{old} occurs in \code{x}, but not if multiple values of \code{old}
occur in \code{x}.
}
\section{Programming notes}{

Factor levels can be replaced through a named list:
\code{f <- addNA(as.factor(c(letters[11:13], NA_character_)))}
\code{levels(f) <- list(c = "k", b = "l", a = "m", h = NA_character_)}
}

\examples{
# All values in 'x' that match any value of 'old' are replaced by 'new'.
replace_vals(x = letters[c(12:13, 13)], old = letters[13:11], new = "b",
             allow_multiple = TRUE)

# If 'allow_multiple' is FALSE, an error is thrown if multiple values of
# 'old' match 'x'.
try(replace_vals(x = letters[c(12:13, 13)], old = letters[13:11], new = "b",
             allow_multiple = FALSE))

# Case-insensitive matching is used if 'ignore_case' is TRUE
replace_vals(x = "A", old = "a", new = "b", ignore_case = TRUE)

# Case-sensitive matching is used if 'ignore_case' is FALSE
# warn_absent = TRUE: warn if no match is found
replace_vals(x = "A", old = "a", new = "b", ignore_case = FALSE,
             warn_absent = TRUE)

# Factor input to 'x'
replace_vals(x = as.factor(c("k", "l", "m")), old = c("k", "l"), new = "b",
             ignore_case = TRUE, allow_multiple = TRUE, warn_absent = TRUE)

replace_vals(x = as.factor(c("m", "l", "k", "M", "L", "K", "m")),
             old = c("M", "k"), new = "x", ignore_case = FALSE,
             allow_multiple = TRUE)

}
\seealso{
Other functions to modify character vectors: 
\code{\link{as.numeric_safe}()},
\code{\link{paste_quoted}()},
\code{\link{reorder_cols}()},
\code{\link{vect_to_char}()},
\code{\link{wrap_text}()}

Other functions to modify factors: 
\code{\link{as.numeric_safe}()},
\code{\link{reorder_levels}()}
}
\concept{functions to modify character vectors}
\concept{functions to modify factors}
\concept{functions to replace values}
