% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_levels.R
\name{round_levels}
\alias{round_levels}
\title{Round levels}
\usage{
round_levels(
  x,
  level_order = NULL,
  digits = 4L,
  type = c("selective", "expanded")
)
}
\arguments{
\item{x}{\link{factor} or \link{character} vector containing
\link[=as.numeric_safe]{numerish values} to be rounded, see \code{Details}.}

\item{level_order}{\code{NULL} to sort levels on increasing numerical value, or a
a vector indicating the desired order of the factor \link{levels}.}

\item{digits}{the minimum number of \link[=signif]{significant digits} to round
to (rounded to the nearest integer in the range from 1 to 22), see \code{Details}.}

\item{type}{\link[checkinput:is_character]{character string} \code{"selective"} or
\code{"expanded"} to indicate the type of rounding to be used, see \code{Details}.}
}
\value{
\code{x} with rounded factor levels.
}
\description{
Round numeric values of factor levels, e.g., to create facet labels that fit
better.
}
\details{
Arguments \code{digits} and \code{type} are passed to \code{\link[=signif_custom]{signif_custom()}}.

Levels of \code{x} that are not present in its values will be silently dropped.

Values in \code{x} that do not occur in \code{level_order} will be added to
\code{level_order}, with a warning. Values in \code{level_order} that do not occur in
\code{x} will be dropped, with a warning.
}
\examples{
round_levels(x = factor(c(2, 2.2, 2.8)), level_order = c("2.2", "2.8", "2"))
round_levels(x = factor(c(2, 2.2, 2.8)), level_order = c("2.2", "2.8", "2.0"))
round_levels(x = factor(c(2, 2.2, 2.8)), level_order = NULL)

}
\seealso{
Other functions to check equality:
\code{\link[=are_equal]{are_equal()}},
\code{\link[=check_case]{check_case()}},
\code{\link[=get_file_path]{get_file_path()}},
\code{\link[=not_in]{not_in()}},
\code{\link[=replace_nonalnum]{replace_nonalnum()}},
\code{\link[=replace_vals]{replace_vals()}},
\code{\link[=signif_custom]{signif_custom()}}

Other functions to modify factors:
\code{\link[=as.numeric_safe]{as.numeric_safe()}},
\code{\link{reexports}},
\code{\link[=reorder_levels]{reorder_levels()}},
\code{\link[=replace_vals]{replace_vals()}},
\code{\link[=vect_to_char]{vect_to_char()}}
}
\concept{functions to check equality}
\concept{functions to modify factors}
