% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_text.R
\name{signal_text}
\alias{signal_text}
\title{Signal a text to the user}
\usage{
signal_text(
  text,
  signal = c("error", "warning", "message", "quiet"),
  origin = character(0)
)
}
\arguments{
\item{text}{Vector with text to be signalled, coerced to character by
\code{\link[=vect_to_char]{vect_to_char()}}.}

\item{signal}{Character string indicating the type of signal to be used:
\code{"error"} to throw an \link[=stop]{error}, \code{"warning"} to issue a \link{warning},
\code{"message"} to show a \link{message}, or \code{"quiet"} to be quiet.}

\item{origin}{Character string pasted behind \code{text} giving the origin of the
message. Ignored if it is \code{character(0)}.}
}
\value{
\code{text}, returned \link[=invisible]{invisibly}.
}
\description{
Signal a text to the user through an error, warning, or message.
}
\details{
\code{text} is coerced to a character vector using \code{\link[=vect_to_char]{vect_to_char()}}, which treats
input with length larger than one better than \link{message} etc. that use
\code{\link[=paste0]{paste0()}}, see the \code{Examples}. Call \code{\link[=vect_to_char]{vect_to_char()}} beforehand on \code{text} to
control rounding and wrapping.
}
\section{Side effects}{

\code{text} is signalled through an \link[=stop]{error}, \link{warning}, \link{message}, or
quietly, depending on argument \code{signal}.
}

\examples{
test_text <- c("Some text", "Some other text")
try(signal_text(text = test_text, signal = "error"))
signal_text(text = test_text, signal = "message")
signal_text(text = test_text, signal = "quiet")

# signal_text() handles input to 'text' with length larger than one more
# sensible than message():
test_numbers <- 11:13
signal_text(text = test_numbers, signal = "message")
message(test_numbers)

# Call vect_to_char() beforehand on 'text' to control rounding of numbers and
# wrapping of text:
message(test_numbers / 7)
signal_text(text = test_numbers / 7, signal = "message")
signal_text(text = vect_to_char(x = test_numbers / 7, signif = 4, width = 15),
            signal = "message")

}
\seealso{
\code{\link[=wrap_text]{wrap_text()}}
}
