% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signif_custom.R
\name{signif_custom}
\alias{signif_custom}
\title{Round numbers}
\usage{
signif_custom(x, digits = 3L, type = c("selective", "expanded"))
}
\arguments{
\item{x}{\link[checkinput:all_numbers]{numeric vector}, \link{matrix} or
\link{data.frame}, see \code{Details}.}

\item{digits}{the \strong{minimum} number of \link[=signif]{significant digits} to
round to (see \code{Details}) or \code{Inf} to not round values.}

\item{type}{\link[checkinput:is_character]{character string} \code{"selective"} or
\code{"expanded"} to indicate the type of rounding to be used, see \code{Details}.}
}
\value{
\code{x} with rounded values, see \code{Details}.
}
\description{
Round numbers to a specified minimum number of significant digits while
ensuring that the integer part is not rounded.
}
\details{
\link[=signif]{Rounding} numbers to a specified number of significant digits can
lead to values ending in rounded zeros, e.g., \code{"14300"} instead of \code{"14286"}
when rounding \code{(1e5)/7} to 3 significant digits. To prevent this,
\code{signif_custom()} treats the value in argument \code{digits} as a \strong{minimum}
value and rounds to more significant digits if that is needed to not round
the integer part of numbers.

If \code{type} is \code{"selective"}, the required value of \code{digits} is determined for
each value individually, such that the number of significant digits might
differ if \code{x} has length larger than one. If type is \code{"expanded"}, the value
of \code{digits} is increased to the same value for all numbers in a vector or
column.

\link[=Inf]{Infinite values} are is returned unchanged. \code{NA_integer_} and
\code{NA_real_} are both returned as \link[=NA]{NA_real_}. \code{signif_custom()} also
handles \link[=matrix]{matrices} and \link[=data.frame]{dataframes}, rounding the numeric
columns with rounding of type \code{expanded} type on a per-column basis. It does
\strong{not} handle \link[=factor]{factors}, use \code{\link[=as.numeric_safe]{as.numeric_safe()}} on \code{x} or use
\code{round_levels(x = x, level_order = levels(x), digits = digits, type = type)}.
}
\examples{
x1 <- (1e5)/7
signif(x = x1, digits = 3) # returns 14300
signif_custom(x1, digits = 3) # returns 14286 to not round the integer part

x2 <- c(10^c(-1, 5)/7)
signif(x = x2, digits = 3) # returns c(1.43e-02, 1.43e+04)
# Increase the number of digits only for the
# second number, returning c(0.0143, 14286.0000)
signif_custom(x2, digits = 3, type = "selective")
# Increase the number of digits for both
# numbers, returning c(0.014286, 14286.0000)
signif_custom(x2, digits = 3, type = "expanded")

x4 <- c(-0.1, 1, -1e4, 1e5) / 7
signif(x = x4, digits = 3)
# returns c(-0.0143,   0.143,   -1430,   14300)
signif_custom(x = x4, digits = 3, type = "selective")
# returns c(-0.0143,   0.1430,  -1429.0, 14286)
signif_custom(x = x4, digits = 3, type = "expanded")
# returns c(-0.014286, 0.14286, -1428.6, 14286)

x_df <- data.frame(a = 1:3, b = letters[11:13], c = c(1e5, 1e3, 1)/7, d = pi)
signif_custom(x_df, type = "selective")
signif_custom(x_df, type = "expanded")

}
\seealso{
\code{\link[=round]{round()}} to round to a specified number of decimal places; \code{\link[=signif]{signif()}} to
round to a specified number of significant digits; \code{\link[=zapsmall]{zapsmall()}} to put small
values to zero; \code{\link[=formatC]{formatC()}} for other ways to format numbers;
\code{\link[=round_levels]{round_levels()}} to round \link{factor} levels.

Other functions to check equality:
\code{\link[=are_equal]{are_equal()}},
\code{\link[=check_case]{check_case()}},
\code{\link[=get_file_path]{get_file_path()}},
\code{\link[=not_in]{not_in()}},
\code{\link[=replace_nonalnum]{replace_nonalnum()}},
\code{\link[=replace_vals]{replace_vals()}},
\code{\link[=round_levels]{round_levels()}}
}
\concept{functions to check equality}
