% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signif_custom.R
\name{signif_custom}
\alias{signif_custom}
\title{Round numbers}
\usage{
signif_custom(x, digits = 3L, type = c("selective", "expanded"))
}
\arguments{
\item{x}{\link[=is.numeric]{numeric vector}, \link{matrix} or \link{data.frame}, see
\code{Details}.}

\item{digits}{the \strong{minimum} number of \link[=signif]{significant digits} to
round to (see \code{Details}) or \code{Inf} to not round values.}

\item{type}{\link[checkinput:is_character]{character string} \code{"selective"} or
\code{"expanded"} indicating the type of rounding to be used, see \code{Details}.}
}
\value{
\code{x} with values rounded according to \code{digits} and \code{type}.
}
\description{
Round numbers to a specified minimum number of \link[=signif]{significant digits}
while ensuring the integer part is not rounded. This prevents values from
ending in rounded zeros, e.g., \code{14300} instead of \code{14286} when rounding
\code{1e5 / 7} to three significant digits.
}
\details{
\link[=Inf]{Infinite values} are returned unchanged. \code{NA_integer_} and \code{NA_real_}
are returned as \link[=NA]{NA_real_}. \link[=matrix]{matrices} and
\link[=data.frame]{dataframes} are handled by rounding values of numeric columns.
\link[=is.numeric]{Non-numeric} (e.g., \link{factor}) \code{x} is \strong{not} supported, use
\code{\link[=as.numeric_safe]{as.numeric_safe()}} beforehand on \code{x} if appropriate.

Numeric values are rounded according to arguments \code{digits} and \code{type}. Values
are \strong{not} rounded if \code{digits} is \link{Inf}. Otherwise, the value of \code{digits}
rounded to the nearest positive integer is used as the \strong{minimum} number of
\link[=signif]{significant digits} to round numeric values in \code{x} to while
ensuring their integer parts are not rounded.

If \code{type} is \code{"selective"}, values are individually rounded to more
significant digits than \code{digits} if that is needed to not round their integer
part, such that the number of significant digits might differ between values
in a vector or in a column. If \code{type} is \code{"expanded"}, all values in a vector
or a in a column are rounded to the same number of significant digits.
}
\examples{
x1 <- (1e5) / 7
signif(x = x1, digits = 3) # returns 14300
signif_custom(x1, digits = 3) # returns 14286 to not round the integer part

x2 <- 10^c(-1, 5) / 7
signif(x = x2, digits = 3) # returns c(1.43e-02, 1.43e+04)
# Increase the number of digits only for the
# second number, returning c(0.0143, 14286.0000)
signif_custom(x2, digits = 3, type = "selective")
# Increase the number of digits for both
# numbers, returning c(0.014286, 14286.0000)
signif_custom(x2, digits = 3, type = "expanded")

x4 <- c(-0.1, 1, -1e4, 1e5) / 7
signif(x = x4, digits = 3)
# returns c(-0.0143,   0.143,   -1430,   14300)
signif_custom(x = x4, digits = 3, type = "selective")
# returns c(-0.0143,   0.1430,  -1429.0, 14286)
signif_custom(x = x4, digits = 3, type = "expanded")
# returns c(-0.014286, 0.14286, -1428.6, 14286)

x_df <- data.frame(a = 1:3, b = letters[11:13], c = c(1e5, 1e3, 1) / 7, d = pi)
signif_custom(x_df, type = "selective")
signif_custom(x_df, type = "expanded")

}
\seealso{
\code{\link[=round]{round()}} to round to a specified number of decimal places; \code{\link[=signif]{signif()}} to
round to a specified number of significant digits; \code{\link[=zapsmall]{zapsmall()}} to put small
values to zero; \code{\link[=formatC]{formatC()}} for other ways to format numbers;
\code{\link[=round_factor]{round_factor()}} to round \link[=factor]{factors}.

Other functions to check equality:
\code{\link[=are_equal]{are_equal()}},
\code{\link[=check_case]{check_case()}},
\code{\link[=get_file_path]{get_file_path()}},
\code{\link[=not_in]{not_in()}},
\code{\link[=replace_nonalnum]{replace_nonalnum()}},
\code{\link[=replace_vals]{replace_vals()}},
\code{\link[=round_factor]{round_factor()}}
}
\concept{functions to check equality}
