% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpaste_unquote.R
\name{unpaste_unquote}
\alias{unpaste_unquote}
\title{Create a character vector from a string with quotation marks}
\usage{
unpaste_unquote(x, collapse = c(", ", "; "), quotemarks = c("'", "\\""))
}
\arguments{
\item{x}{A character string}

\item{collapse}{Character vector with elements that were used to collapse
values when \code{x} was created, and are now used to \link[=strsplit]{split} \code{x} on.
Can be \code{character(0)} to leave \code{x} as a character string.}

\item{quotemarks}{Character vector with quotation marks to be removed from
\code{x}. Can be \code{character(0)} to not remove any quotation marks.}
}
\value{
\code{x} without the quotation marks and split into a vector.
}
\description{
Create a character vector from a string with quotation marks
}
\details{
\code{unpaste_unquote()} is not the exact reverse of \code{\link[=paste_quoted]{paste_quoted()}}: it does
\emph{not} restore zero-length elements to their original value, e.g., \code{"'NULL'"}
to \code{NULL}, or \code{"'character(0)'"} to \code{character(0)}.
}
\examples{
x <- "'ff', 'gG', 'HH'"
unpaste_unquote(x = x, collapse = ", ", quotemarks = "'")
unpaste_unquote(x = x, collapse = ", ", quotemarks = "\"")
unpaste_unquote(x = x, collapse = character(0), quotemarks = c("'", "\""))
unpaste_unquote(x = x, collapse = character(0), quotemarks = character(0))

}
\seealso{
\code{\link[=paste_quoted]{paste_quoted()}} for the opposite of \code{unpaste_unquote()}

Other functions to modify character vectors: 
\code{\link{as.numeric_safe}()},
\code{\link{file_path_sans_ext}()},
\code{\link{paste_quoted}()},
\code{\link{reorder_cols}()},
\code{\link{replace_vals}()},
\code{\link{replace_vals_old}()},
\code{\link{vect_to_char}()},
\code{\link{wrap_text}()}
}
\concept{functions to modify character vectors}
