% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vect_to_char.R
\name{vect_to_char}
\alias{vect_to_char}
\title{Convert a vector to a character string}
\usage{
vect_to_char(
  x,
  signif = 3L,
  width = Inf,
  sep = ": ",
  collapse = ", ",
  ignore_newlines = TRUE
)
}
\arguments{
\item{x}{A vector, \link{factor}, \code{NULL}, or a non-dataframe \link{list} (unlisted
through \code{unlist(x, use.names = TRUE)}, with a warning).}

\item{signif}{Positive number of length one, rounded to the nearest positive
integer indicating the number of significant digits to round numeric \code{x} to.}

\item{width}{\link[checkinput:is_natural]{natural number} giving the maximum
line width (in characters)
after wrapping. Can be \code{Inf} to not wrap text.}

\item{sep}{\link[checkinput:is_character]{character string} used to separate
the names and the values. Ignored if \code{x} does not have names.}

\item{collapse}{\link[checkinput:is_character]{character string} to collapse
values into a single character string, or \code{NULL} to return each value as an
element of a character vector.}

\item{ignore_newlines}{\code{TRUE} or \code{FALSE}: should newlines in \code{x} be replaced
by blank characters?}
}
\value{
A character vector or character string with the names and values in \code{x}, with
values of numeric \code{x} rounded to \code{signif} \link[=signif]{significant} digits (but
at least the integer part displayed compeletely, see \code{\link[=signif_custom]{signif_custom()}}),
wrapped to \code{width} characters.

If \code{collapse} is \code{NULL}, a \link[checkinput:all_characters]{character vector}
with the same elements as \code{x}
is returned, wrapping on a per-element basis. If \code{collapse} is not \code{NULL},
the name-value pairs are separated by \code{collapse}, thus returning a
\link[checkinput:is_character]{character string}.

See \code{Details} on handling of some special values.
}
\description{
Convert a vector to a character string while preserving names and rounding
numeric values.
}
\details{
Some values are handled specially to better distinguish different values than
\code{\link[=message]{message()}} etc. that use \code{\link[=paste0]{paste0()}}:
\itemize{
\item \code{NULL} is returned as \code{"NULL"}
\item non-\code{NULL} zero-length objects are returned as \code{"<type>(0)"} (e.g.,
\code{"logical(0)"}; for \code{numeric(0)} this is \code{"double(0)"})
\item \code{""} is returned as \code{""}
\item logical \code{NA} is returned as \code{"NA"}
\item non-logical \code{NA} is returned as \code{"NA_<type>_"} (e.g., \code{"NA_character_"};
for \code{NA_real_} this is \code{"NA_double_"};
for \link[=factor]{factors} this is \code{"NA_character_"} because \code{vect_to_char()}
converts factors to characters).
}
}
\section{Programming notes}{

To get a cross-tabulation of \code{x} into a character string, one can use
\code{paste0(vect_to_char(c(table(x)), sep = " (", collapse =  "), "), ")")},
see the last \code{Example}.
}

\examples{
x <- 1:3
names(x) <- letters[x]
vect_to_char(x = x) # "a: 1, b: 2, c: 3"
vect_to_char(x = x, collapse = NULL) # c("a: 1", "b: 2", "c: 3")
vect_to_char(x = unname(x)) # "1, 2, 3"
y <- x / 7
vect_to_char(x = y, signif = 7) # "a: 0.1428571, b: 0.2857143, c: 0.4285714"
vect_to_char(x = y, signif = 2, sep = " = ", collapse = " and ", width = 15)
# "a = 0.14 and b\n= 0.29 and c =\n0.43"
vect_to_char(x = 1e3/7, signif = 2) # not 140: completely display integer part

x_char <- c(a = "abc", b = "def", c = "this is text")
vect_to_char(x = x_char) # "a: abc, b: def, c: this is text"
vect_to_char(x = unname(x_char)) # "abc, def, this is some text"

# Nicer handling of zero-length input
vect_to_char(logical(0)) # "logical(0)"
message(logical(0)) # <empty>
message(vect_to_char(logical(0))) # logical(0)

# Using vect_to_char() to get a frequency table
x <- 1:10
names(x) <- letters[x]
y <- 5:15
names(y) <- letters[y]
x <- c(x, y)
paste0(vect_to_char(c(table(x)), sep = " (", collapse =  "), "), ")")

}
\seealso{
\code{\link[=toString]{toString()}} which is shorthand for \code{paste(x, collapse = ", ")} and can be
used instead of \code{vect_to_char()} if names of \code{x} can be removed;
\code{vignette("type_coercion", package = "checkinput")} and
\code{help("is_zerolength", package = "checkinput")} for a discussion of some
issues with type conversion and zero-length input when \link[=c]{combining}
objects into a vector.

Other functions to convert types:
\code{\link[=as.numeric_safe]{as.numeric_safe()}},
\code{\link{reexports}},
\code{\link[=reorder_levels]{reorder_levels()}}

Other functions to modify character vectors:
\code{\link[=as.numeric_safe]{as.numeric_safe()}},
\code{\link{reexports}},
\code{\link[=replace_vals]{replace_vals()}},
\code{\link[=signal_text]{signal_text()}},
\code{\link[=unpaste_unquote]{unpaste_unquote()}},
\code{\link[=wrap_text]{wrap_text()}}

Other functions to modify factors:
\code{\link[=as.numeric_safe]{as.numeric_safe()}},
\code{\link{reexports}},
\code{\link[=reorder_levels]{reorder_levels()}},
\code{\link[=replace_vals]{replace_vals()}},
\code{\link[=round_levels]{round_levels()}}
}
\concept{functions to convert types}
\concept{functions to modify character vectors}
\concept{functions to modify factors}
