% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vect_to_char.R
\name{vect_to_char}
\alias{vect_to_char}
\title{Convert a vector to a character string}
\usage{
vect_to_char(x, signif = 3L, sep = ": ", collapse = ", ", width = Inf)
}
\arguments{
\item{x}{A vector, \link{factor}, non-dataframe \link{list}, or \code{NULL}.}

\item{signif}{Positive number of length one, rounded to the nearest positive
integer indicating the number of significant digits to round numeric \code{x} to.}

\item{sep}{Character string of length one used to separate the names and the
values. Not used if \code{x} does not have names.}

\item{collapse}{Character string of length one to collapse values into a
single character string, or \code{NULL} to return each value as an element of a
character vector.}

\item{width}{A positive number (rounded to a whole number) giving the maximum
line width (in characters) after wrapping. Can be \code{Inf} to not wrap text.}
}
\value{
A character string containing the names and rounded values in \code{x}, separating
name-value pairs by \code{collapse}. If \code{collapse} is \code{NULL}, a character \emph{vector}
with the same elements as \code{x} is returned. The output is wrapped to \code{width}
characters.
}
\description{
Convert a vector to a character string, preserving names, rounding numeric
values, returning \code{NULL} as \code{"'NULL'"} and returning other zero-length
objects as \code{"'<class>(0)'"}, e.g., \code{"'logical(0)'"}.
}
\details{
Numeric \code{x} is rounded to \code{signif} digits.

\code{NULL} is returned as \code{"'NULL'"}, other zero-length objects are returned as
\code{"'<class>(0)'"}, e.g., \code{"'logical(0)'"}.

The output is wrapped to \code{width} characters if \code{width} is not \code{NULL}. This
occurs on a per-element basis if a character vector is returned because
\code{collapse} is \code{NULL}.
}
\section{Wishlist}{

Write a similar function for dataframes or matrices, with the name of the
relevant columns only once.

Implement customized rounding using \code{signif_custom()}, e.g.,
\code{x <- signif_custom(x = x, digits = signif, type = type)}.

See also:
\itemize{
\item wrMisc::convToNum()
\item num_as_char() in https://github.com/hneth/ds4psy/blob/master/R/num_util_fun.R
\item drop_sci_note() in https://github.com/trinker/textclean/blob/master/R/utils.R
}
}

\section{Programming note}{

To get a cross-tabulation of \code{x} into a character string, one could use
\code{paste0(vect_to_char(c(table(z)), sep = " (", collapse =  "), "), ")")},
see the last \code{Example}.
}

\examples{
x <- 1:3
names(x) <- letters[x]
vect_to_char(x = x) # "a: 1, b: 2, c: 3"
vect_to_char(x = x, collapse = NULL) # c("a: 1", "b: 2", "c: 3")
vect_to_char(x = unname(x)) # "1, 2, 3"
y <- x / 7
vect_to_char(x = y, signif = 7) # "a: 0.1428571, b: 0.2857143, c: 0.4285714"
vect_to_char(x = y, signif = 2, sep = " = ", collapse = " and ", width = 15)
# "a = 0.14 and b\n= 0.29 and c =\n0.43"

x_char <- c(a = "abc", b = "def", c = "this is text")
vect_to_char(x = x_char) # "a: abc, b: def, c: this is text"
vect_to_char(x = unname(x_char)) # "abc, def, this is some text"

# Using vect_to_char() to get a frequency table
x <- 1:10
names(x) <- letters[x]
y <- 5:15
names(y) <- letters[y]
z <- c(x, y)
paste0(vect_to_char(c(table(z)), sep = " (", collapse =  "), "), ")")

}
\seealso{
\code{\link[=toString]{toString()}} which can be used if names of \code{x} can be removed.

Other functions to modify character vectors: 
\code{\link{as.numeric_safe}()},
\code{\link{paste_quoted}()},
\code{\link{wrap_text}()}
}
\concept{functions to modify character vectors}
