% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vect_to_char.R
\name{vect_to_char}
\alias{vect_to_char}
\title{Convert a vector to a character string}
\usage{
vect_to_char(x, signif = 3L, width = Inf, sep = ": ", collapse = ", ")
}
\arguments{
\item{x}{A vector, \link{factor}, non-dataframe \link{list}, or \code{NULL}.}

\item{signif}{Positive number of length one, rounded to the nearest positive
integer indicating the number of significant digits to round numeric \code{x} to.}

\item{width}{A positive number giving the maximum line width (in characters)
after wrapping. Can be \code{Inf} to not wrap text.}

\item{sep}{Character string of length one used to separate the names and the
values. Ignored if \code{x} does not have names.}

\item{collapse}{Character string of length one to collapse values into a
single character string, or \code{NULL} to return each value as an element of a
character vector.}
}
\value{
The names and values in \code{x}, with values of numeric \code{x} rounded to \code{signif}
\link[=signif]{significant} digits, wrapped to \code{width} characters.

If \code{collapse} is \code{NULL}, a character \emph{vector} with the same elements as \code{x}
is returned, wrapping on a per-element basis. If \code{collapse} is not \code{NULL},
the name-value pairs are separated by \code{collapse}, thus returning a character
\emph{string}.

\code{NULL} is returned as \code{"'NULL'"}, other zero-length objects are returned as
\code{"'<class>(0)'"}, e.g., \code{"'logical(0)'"}.
}
\description{
Convert a vector to a character string, preserving names, rounding numeric
values, returning \code{NULL} as \code{"'NULL'"} and other zero-length objects as
\code{"'<class>(0)'"}, e.g., \code{"'logical(0)'"}.
}
\section{Programming note}{

To get a cross-tabulation of \code{x} into a character string, one can use
\code{paste0(vect_to_char(c(table(x)), sep = " (", collapse =  "), "), ")")},
see the last \code{Example}.
}

\examples{
x <- 1:3
names(x) <- letters[x]
vect_to_char(x = x) # "a: 1, b: 2, c: 3"
vect_to_char(x = x, collapse = NULL) # c("a: 1", "b: 2", "c: 3")
vect_to_char(x = unname(x)) # "1, 2, 3"
y <- x / 7
vect_to_char(x = y, signif = 7) # "a: 0.1428571, b: 0.2857143, c: 0.4285714"
vect_to_char(x = y, signif = 2, sep = " = ", collapse = " and ", width = 15)
# "a = 0.14 and b\n= 0.29 and c =\n0.43"

x_char <- c(a = "abc", b = "def", c = "this is text")
vect_to_char(x = x_char) # "a: abc, b: def, c: this is text"
vect_to_char(x = unname(x_char)) # "abc, def, this is some text"

# Using vect_to_char() to get a frequency table
x <- 1:10
names(x) <- letters[x]
y <- 5:15
names(y) <- letters[y]
x <- c(x, y)
paste0(vect_to_char(c(table(x)), sep = " (", collapse =  "), "), ")")

}
\seealso{
\code{\link[=toString]{toString()}} which can be used if names of \code{x} can be removed.

Other functions to modify character vectors: 
\code{\link{as.numeric_safe}()},
\code{\link{paste_quoted}()},
\code{\link{wrap_text}()}
}
\concept{functions to modify character vectors}
