% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_text.R
\name{wrap_text}
\alias{wrap_text}
\title{Wrap text}
\usage{
wrap_text(x, width = 80L, ignore_newlines = TRUE, warn_length = FALSE)
}
\arguments{
\item{x}{\link[checkinput:all_characters]{character vector} to be wrapped.}

\item{width}{\link[checkinput:is_natural]{natural number} giving the maximum
line width (in characters)
after wrapping. Can be \code{Inf} to not wrap text.}

\item{ignore_newlines}{\code{TRUE} or \code{FALSE}: should newlines in \code{x} be replaced
by blank characters?}

\item{warn_length}{\code{TRUE} or \code{FALSE}: warn if a fragment exceeds \code{width}?}
}
\value{
A \link[checkinput:is_character]{character string} containing \code{x} wrapped to a
maximum of \code{width} characters, with newlines inserted at blank characters. A
warning is issued if \code{warn_length} is \code{TRUE} and the width of a fragment in
the output exceeds \code{width}: this occurs if a stretch of characters longer
than \code{width} occurs without a blank character to wrap at.
}
\description{
Wrap text at blank characters to achieve a maximum line width.
}
\details{
\code{character(0)} input to \code{x} is returned unchanged, with a warning.

\code{x} of length larger than one is pasted into a single string, separating the
parts by blank characters.

Consecutive white space is collapsed into a single blank character, except
for double spaces after periods, question marks and exclamation marks (as
documented in the section \code{Details} of \code{\link[=strwrap]{strwrap()}}).

Leading and trailing newlines are collapsed into a single blank character if
\code{ignore_newlines} is \code{TRUE} and are retained if \code{ignore_newlines} is \code{FALSE}.
}
\section{Programming notes}{

The call \code{wrap_text(x, width)} can be replaced by
\code{paste0(strwrap(x, width + 1L), collapse = "\\n")} if \code{x} has length one and
\code{ignore_newlines} is \code{TRUE}.

Using \code{wrap_text()} on \code{x} of variable length, e.g., in the text of warnings
that report a file path or a user-provided variable name, makes the location
of newlines unpredictable and thus difficult to test reliably. To circumvent
this, put the constant part at the front of the message and hardcode newlines
using \verb{\\n}.
}

\section{Notes}{

The output is printed as a string with newlines represented as \verb{\\n}. Use
\code{\link[=cat]{cat()}} on the output to print it in the way it is formatted in messages.

Argument \code{width} in \code{wrap_text()} indicates the maximum width of text after
wrapping, i.e., the width \strong{after} which text should be wrapped. In
contrast, argument \code{width} in \code{\link[=strwrap]{strwrap()}} indicates the width \strong{at} which
text should be wrapped.
}

\examples{
example_text <- "A piece\nof text that you want to wrap over multiple lines"
cat(wrap_text(example_text,
              width = 20, ignore_newlines = TRUE))
cat(wrap_text(example_text,
              width = 20, ignore_newlines = FALSE))

}
\seealso{
\code{\link[=cat]{cat()}}; \code{\link[=paste]{paste()}}; \code{\link[=strtrim]{strtrim()}}; \code{\link[=strwrap]{strwrap()}}

Other functions to modify character vectors:
\code{\link[=as.numeric_safe]{as.numeric_safe()}},
\code{\link{reexports}},
\code{\link[=replace_vals]{replace_vals()}},
\code{\link[=signal_text]{signal_text()}},
\code{\link[=unpaste_unquote]{unpaste_unquote()}},
\code{\link[=vect_to_char]{vect_to_char()}}
}
\concept{functions to modify character vectors}
