% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_text.R
\name{wrap_text}
\alias{wrap_text}
\title{Wrap text}
\usage{
wrap_text(x, width = 80L, ignore_newlines = TRUE)
}
\arguments{
\item{x}{Character vector to be wrapped.}

\item{width}{A positive number giving the maximum line width (in characters)
after wrapping. Can be \code{Inf} to not wrap text.}

\item{ignore_newlines}{\code{TRUE} or \code{FALSE}: should newlines in \code{x} be replaced
by a blank character? Newlines at the end of \code{x} are always removed.}
}
\value{
A string containing \code{x} wrapped to a maximum of \code{width} characters, with
newlines inserted at blank characters. A warning is issued if the width of a
fragment in the output exceeds \code{width}: this occurs if a stretch of
characters longer than \code{width} occurs without a blank character to wrap at.
}
\description{
Wrap text at blank characters to achieve a maximum line width.
}
\details{
\code{x} of length larger than one is pasted into a single string, separating the
parts by blank characters.

Leading white space in \code{x} is completely removed.

Consecutive white space in \code{x} is collapsed into a single blank character,
except for double spaces after periods, question marks and exclamation marks
(as documented in the section \code{Details} of \code{\link[=strwrap]{strwrap()}}).
}
\note{
The output is printed as a string with newlines represented as \verb{\\n}. Use
\code{\link[=cat]{cat()}} on the output to print it in the way it is formatted in messages.

Argument \code{width} in \code{wrap_text()} indicates the maximum width of text after
wrapping, i.e., the width \emph{after} which text should be wrapped. In contrast,
argument \code{width} in \code{\link[=strwrap]{strwrap()}} indicates the width \emph{at} which text should be
wrapped.
}
\section{Programming note}{

The call \code{wrap_text(x, width)} can be replaced by
\code{paste0(strwrap(x, width + 1L), collapse = "\\n")} if \code{x} has length one and
\code{ignore_newlines} is \code{TRUE}.
}

\examples{
example_text <- "A piece\nof text that you want to wrap over multiple lines"
cat(wrap_text(example_text,
              width = 20, ignore_newlines = TRUE))
cat(wrap_text(example_text,
              width = 20, ignore_newlines = FALSE))

}
\seealso{
\code{\link[=cat]{cat()}}; \code{\link[=paste]{paste()}}; \code{\link[=strwrap]{strwrap()}}; \code{bbmle:::strwrapx()}.

Other functions to modify character vectors: 
\code{\link{as.numeric_safe}()},
\code{\link{paste_quoted}()},
\code{\link{vect_to_char}()}
}
\concept{functions to modify character vectors}
