% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_text.R
\name{wrap_text}
\alias{wrap_text}
\title{Wrap text}
\usage{
wrap_text(x, width = 80L, ignore_newlines = TRUE)
}
\arguments{
\item{x}{Character vector to be wrapped.}

\item{width}{A positive number (rounded to a whole number) giving the maximum
line width (in characters) after wrapping. Can be \code{Inf} to not wrap text.}

\item{ignore_newlines}{\code{TRUE} or \code{FALSE}: should newlines in \code{x} be replaced
by a space character? If \code{FALSE}, newlines are respected, although a newline
at the end of \code{x} is always removed.}
}
\value{
\code{x} with newlines inserted at blank characters to wrap \code{x} to
a maximum of \code{width} characters. \code{x} of length larger than one is pasted into
a single string, separating the parts by spaces. Consecutive white space in
\code{x} is collapsed into a single blank character, except for double spaces
after periods, question marks and exclamation marks (as documented in the
section \code{Details} of \code{\link[base:strwrap]{base::strwrap()}}). Leading white space in \code{x} is
completely removed.
}
\description{
Wrap text at blank characters to achieve a maximum line width.
}
\note{
A warning is issued if the width of a fragment in the output exceeds \code{width}.

The output is printed as a string with newlines represented as \verb{\\n}. Use
\code{\link[base:cat]{base::cat()}} on the output to print it in the way it is formatted in messages.

Argument \code{width} in \code{wrap_text()} indicates the maximum width of text after
wrapping, i.e., the width \emph{after} which text should be wrapped. In contrast,
argument \code{width} in \code{\link[base:strwrap]{base::strwrap()}} indicates the width \emph{at} which text
should be wrapped.
}
\section{Programming note}{

The call \code{wrap_text(x, width)} can be replaced by
\code{paste0(strwrap(x, width + 1L), collapse = "\\n")} if \code{x} has length one and
\code{ignore_newlines} is \code{TRUE}.
}

\examples{
cat(wrap_text("A piece\nof text that you want to wrap over multiple lines",
              width = 20, ignore_newlines = TRUE))
cat(wrap_text("A piece\nof text that you want to wrap over multiple lines",
              width = 20, ignore_newlines = FALSE))

}
\seealso{
\code{\link[base:cat]{base::cat()}} \code{\link[base:paste]{base::paste()}} \code{\link[base:strwrap]{base::strwrap()}}

Other functions to modify character vectors: 
\code{\link{as.numeric_safe}()},
\code{\link{numvect_to_char}()},
\code{\link{paste_quoted}()}
}
\concept{functions to modify character vectors}
